refactor(gateway): split session utilities
