+++ src/gateway/session-utils.ts
@@ -1,140 +1,3 @@
-import { expectDefined } from "@openclaw/normalization-core";
-// Gateway session listing and projection helpers.
-// Normalizes persisted session stores into UI/RPC rows without mutating state.
-import {
-  normalizeLowercaseStringOrEmpty,
-  normalizeOptionalString,
-  normalizeOptionalLowercaseString,
-} from "@openclaw/normalization-core/string-coerce";
-import { uniqueStrings } from "@openclaw/normalization-core/string-normalization";
-import type {
-  SessionCreatedActor,
-  SessionsListParams,
-} from "../../packages/gateway-protocol/src/index.js";
-import {
-  readAcpSessionMeta,
-  readAcpSessionMetaForEntry,
-  repairAcpSessionMetaKeyForMigration,
-} from "../acp/runtime/session-meta.js";
-import { resolveModelAgentRuntimeMetadata } from "../agents/agent-runtime-metadata.js";
-import {
-  listAgentIds,
-  resolveAgentConfig,
-  resolveAgentEffectiveModelPrimary,
-  resolveAgentModelFallbacksOverride,
-  resolveAgentWorkspaceDir,
-  resolveDefaultAgentId,
-} from "../agents/agent-scope.js";
-import { lookupContextTokens, resolveContextTokensForModel } from "../agents/context.js";
-import { DEFAULT_CONTEXT_TOKENS, DEFAULT_MODEL, DEFAULT_PROVIDER } from "../agents/defaults.js";
-import { resolveFastModeState } from "../agents/fast-mode.js";
-import { resolveAgentAvatarUrlFromSource } from "../agents/identity-avatar-file.js";
-import {
-  findModelCatalogEntry,
-  modelSupportsInput,
-  type ModelCatalogEntry,
-} from "../agents/model-catalog.js";
-import {
-  inferUniqueProviderFromConfiguredModels,
-  isCliProvider,
-  normalizeStoredOverrideModel,
-  parseModelRef,
-  resolveConfiguredModelRef,
-  resolveDefaultModelForAgent,
-  resolveThinkingDefault,
-} from "../agents/model-selection.js";
-import {
-  resolveSessionModelIdentityRef,
-  resolveSessionModelRef,
-} from "../agents/session-model-ref.js";
-import { resolveSessionRuntimeOverrideForProvider } from "../agents/session-runtime-compat.js";
-import {
-  buildSubagentRunReadIndex,
-  countActiveDescendantRuns,
-  getSessionDisplaySubagentRunByChildSessionKey,
-  getSubagentSessionRuntimeMs,
-  getSubagentSessionStartedAt,
-  isSubagentRunLive,
-  listSubagentRunsForController,
-  resolveSubagentSessionStatus,
-} from "../agents/subagent-registry-read.js";
-import {
-  RECENT_ENDED_SUBAGENT_CHILD_SESSION_MS,
-  shouldKeepSubagentRunChildLink,
-} from "../agents/subagent-run-liveness.js";
-import {
-  concretizeAgentRuntime,
-  resolveEffectiveAgentRuntime,
-} from "../agents/thinking-runtime.js";
-import { insideGitCheckout } from "../agents/worktrees/git.js";
-import { resolveQueueSettings } from "../auto-reply/reply/queue/settings.js";
-import {
-  listThinkingLevelOptions,
-  normalizeThinkLevel,
-  resolveEffectiveResponseUsage,
-  resolveSupportedThinkingLevel,
-} from "../auto-reply/thinking.js";
-import { getRuntimeConfig } from "../config/io.js";
-import { resolveAgentModelFallbackValues } from "../config/model-input.js";
-import {
-  buildGroupDisplayName,
-  buildGroupDisplayTitle,
-  isConfiguredSessionStoreAgentId,
-  isTerminalSessionStatus,
-  resolveAllAgentSessionStoreTargetsSync,
-  resolveAgentMainSessionKey,
-  resolveExistingAgentSessionStoreTargetsSync,
-  resolveFreshSessionTotalTokens,
-  resolveSessionGoalDisplayState,
-  resolveStorePath,
-  type SessionEntry,
-  type SessionStoreTarget,
-  type SessionScope,
-} from "../config/sessions.js";
-import {
-  listSessionEntries as listAccessorSessionEntries,
-  listSessionEntriesReadOnly as listAccessorSessionEntriesReadOnly,
-} from "../config/sessions/session-accessor.js";
-import { sessionEntryForkedFromParent } from "../config/sessions/session-entry-lineage.js";
-import type { OpenClawConfig } from "../config/types.openclaw.js";
-import { projectPluginSessionExtensionsSync } from "../plugins/host-hook-state.js";
-import { withPinnedActivePluginRegistryWorkspaceDir } from "../plugins/runtime-workspace-state.js";
-import {
-  DEFAULT_AGENT_ID,
-  isIncognitoSessionKey,
-  normalizeAgentId,
-  parseAgentSessionKey,
-} from "../routing/session-key.js";
-import { resolveActiveSessionAgentStatus } from "../sessions/session-agent-status.js";
-import { isAcpSessionKey, isCronRunSessionKey } from "../sessions/session-key-utils.js";
-import { resolveNonNegativeNumber } from "../shared/number-coercion.js";
-import { resolveIncognitoOpenClawAgentSqlitePath } from "../state/openclaw-agent-db.js";
-import { getUserProfileListItem } from "../state/user-profiles.js";
-import { truncateUtf16Safe } from "../utils.js";
-import { normalizeSessionDeliveryFields } from "../utils/delivery-context.shared.js";
-import { INTERNAL_MESSAGE_CHANNEL } from "../utils/message-channel-constants.js";
-import type { ModelCostConfig } from "../utils/usage-format.js";
-import { estimateUsageCost, resolveModelCostConfig } from "../utils/usage-format.js";
-import { listGatewayAgentsBasic } from "./agent-list.js";
-import { sessionHasAutomation } from "./session-automation-index.js";
-import { sortAndLimitSessionEntries, type SessionEntryPair } from "./session-list-order.js";
-import {
-  resolveSessionStoreAgentId,
-  resolveSessionStoreKey,
-  resolveStoredSessionKeyForAgentStore,
-} from "./session-store-key.js";
-import {
-  readRecentSessionUsageFromTranscript as readScopedRecentSessionUsageFromTranscript,
-  readSessionTitleFieldsFromTranscriptAsync as readScopedSessionTitleFieldsFromTranscriptAsync,
-  readSessionTitleFieldsFromTranscript as readScopedSessionTitleFieldsFromTranscript,
-} from "./session-transcript-readers.js";
-import type {
-  GatewayAgentRow,
-  GatewaySessionRow,
-  GatewaySessionsDefaults,
-  SessionsListResult,
-} from "./session-utils.types.js";
-
 export {
   resolveSessionHistoryTranscriptPathAsync,
   resolveSessionTranscriptCandidates,
@@ -148,2905 +11,24 @@ export type {
   SessionsPreviewResult,
 } from "./session-utils.types.js";
 export { resolveSessionModelRef } from "../agents/session-model-ref.js";
-
-const DERIVED_TITLE_MAX_LEN = 60;
-
-function formatSessionIdPrefix(sessionId: string, updatedAt?: number | null): string {
-  const prefix = sessionId.slice(0, 8);
-  if (updatedAt && updatedAt > 0) {
-    const d = new Date(updatedAt);
-    const date = d.toISOString().slice(0, 10);
-    return `${prefix} (${date})`;
-  }
-  return prefix;
-}
-
-function truncateTitle(text: string, maxLen: number): string {
-  if (text.length <= maxLen) {
-    return text;
-  }
-  const cut = truncateUtf16Safe(text, maxLen - 1);
-  const lastSpace = cut.lastIndexOf(" ");
-  if (lastSpace > maxLen * 0.6) {
-    return cut.slice(0, lastSpace) + "…";
-  }
-  return cut + "…";
-}
-
-export function deriveSessionTitle(
-  entry: SessionEntry | undefined,
-  firstUserMessage?: string | null,
-): string | undefined {
-  if (!entry) {
-    return undefined;
-  }
-
-  const label = normalizeOptionalString(entry.label);
-  if (label) {
-    return label;
-  }
-
-  if (normalizeOptionalString(entry.displayName)) {
-    return normalizeOptionalString(entry.displayName);
-  }
-
-  if (normalizeOptionalString(entry.subject)) {
-    return normalizeOptionalString(entry.subject);
-  }
-
-  if (firstUserMessage?.trim()) {
-    const normalized = firstUserMessage.replace(/\s+/g, " ").trim();
-    return truncateTitle(normalized, DERIVED_TITLE_MAX_LEN);
-  }
-
-  if (entry.sessionId) {
-    return formatSessionIdPrefix(entry.sessionId, entry.updatedAt);
-  }
-
-  return undefined;
-}
-
-function resolveSessionRuntimeMs(
-  run: { startedAt?: number; endedAt?: number; accumulatedRuntimeMs?: number } | null,
-  now: number,
-) {
-  return getSubagentSessionRuntimeMs(run, now);
-}
-
-function resolvePositiveNumber(value: number | null | undefined): number | undefined {
-  return typeof value === "number" && Number.isFinite(value) && value > 0 ? value : undefined;
-}
-
-function deriveSessionUnread(
-  entry?: Pick<
-    SessionEntry,
-    "lastReadAt" | "markedUnreadAt" | "lastInteractionAt" | "lastActivityAt"
-  >,
-): boolean {
-  return (
-    entry?.markedUnreadAt !== undefined ||
-    (entry?.lastReadAt !== undefined &&
-      Math.max(entry.lastInteractionAt ?? 0, entry.lastActivityAt ?? 0) > entry.lastReadAt)
-  );
-}
-
-type SessionCompactionCheckpointEntry = NonNullable<SessionEntry["compactionCheckpoints"]>[number];
-
-function isProjectableCompactionCheckpoint(
-  value: unknown,
-): value is SessionCompactionCheckpointEntry {
-  if (!value || typeof value !== "object" || Array.isArray(value)) {
-    return false;
-  }
-  const checkpoint = value as {
-    checkpointId?: unknown;
-    createdAt?: unknown;
-    reason?: unknown;
-  };
-  return (
-    Boolean(normalizeOptionalString(checkpoint.checkpointId)) &&
-    typeof checkpoint.createdAt === "number" &&
-    Number.isFinite(checkpoint.createdAt) &&
-    (checkpoint.reason === "manual" ||
-      checkpoint.reason === "auto-threshold" ||
-      checkpoint.reason === "overflow-retry" ||
-      checkpoint.reason === "timeout-retry")
-  );
-}
-
-function resolveProjectableCompactionCheckpoints(
-  entry?: Pick<SessionEntry, "compactionCheckpoints"> | null,
-): SessionCompactionCheckpointEntry[] {
-  const checkpoints = entry?.compactionCheckpoints;
-  if (!Array.isArray(checkpoints) || checkpoints.length === 0) {
-    return [];
-  }
-  return checkpoints.filter(isProjectableCompactionCheckpoint);
-}
-
-function resolveLatestCompactionCheckpoint(
-  checkpoints: readonly SessionCompactionCheckpointEntry[],
-): SessionCompactionCheckpointEntry | undefined {
-  return checkpoints.reduce<SessionCompactionCheckpointEntry | undefined>(
-    (latest, checkpoint) =>
-      !latest || checkpoint.createdAt > latest.createdAt ? checkpoint : latest,
-    undefined,
-  );
-}
-
-function buildCompactionCheckpointPreview(
-  checkpoint: SessionCompactionCheckpointEntry | undefined,
-): GatewaySessionRow["latestCompactionCheckpoint"] {
-  if (!checkpoint) {
-    return undefined;
-  }
-  const checkpointId = normalizeOptionalString(checkpoint.checkpointId);
-  const createdAt = checkpoint.createdAt;
-  const reason = checkpoint.reason;
-  if (!checkpointId || typeof createdAt !== "number" || !Number.isFinite(createdAt)) {
-    return undefined;
-  }
-  if (
-    reason !== "manual" &&
-    reason !== "auto-threshold" &&
-    reason !== "overflow-retry" &&
-    reason !== "timeout-retry"
-  ) {
-    return undefined;
-  }
-  return {
-    checkpointId,
-    createdAt,
-    reason,
-  };
-}
-
-function resolveModelCostConfigCached(
-  provider: string | undefined,
-  model: string | undefined,
-  cfg: OpenClawConfig,
-  rowContext?: SessionListRowContext,
-): ModelCostConfig | undefined {
-  if (!rowContext) {
-    return resolveModelCostConfig({ provider, model, config: cfg });
-  }
-  const key = createSessionRowModelCacheKey(provider, model);
-  if (rowContext.modelCostConfigByModelRef.has(key)) {
-    return rowContext.modelCostConfigByModelRef.get(key);
-  }
-  const value = resolveModelCostConfig({ provider, model, config: cfg });
-  rowContext.modelCostConfigByModelRef.set(key, value);
-  return value;
-}
-
-function resolveEstimatedSessionCostUsd(params: {
-  cfg: OpenClawConfig;
-  provider?: string;
-  model?: string;
-  entry?: Pick<
-    SessionEntry,
-    "estimatedCostUsd" | "inputTokens" | "outputTokens" | "cacheRead" | "cacheWrite"
-  >;
-  explicitCostUsd?: number;
-  rowContext?: SessionListRowContext;
-}): number | undefined {
-  const explicitCostUsd = resolveNonNegativeNumber(
-    params.explicitCostUsd ?? params.entry?.estimatedCostUsd,
-  );
-  if (explicitCostUsd !== undefined) {
-    return explicitCostUsd;
-  }
-  const input = resolvePositiveNumber(params.entry?.inputTokens);
-  const output = resolvePositiveNumber(params.entry?.outputTokens);
-  const cacheRead = resolvePositiveNumber(params.entry?.cacheRead);
-  const cacheWrite = resolvePositiveNumber(params.entry?.cacheWrite);
-  if (
-    input === undefined &&
-    output === undefined &&
-    cacheRead === undefined &&
-    cacheWrite === undefined
-  ) {
-    return undefined;
-  }
-  const cost = resolveModelCostConfigCached(
-    params.provider,
-    params.model,
-    params.cfg,
-    params.rowContext,
-  );
-  if (!cost) {
-    return undefined;
-  }
-  const estimated = estimateUsageCost({
-    usage: {
-      ...(input !== undefined ? { input } : {}),
-      ...(output !== undefined ? { output } : {}),
-      ...(cacheRead !== undefined ? { cacheRead } : {}),
-      ...(cacheWrite !== undefined ? { cacheWrite } : {}),
-    },
-    cost,
-  });
-  return resolveNonNegativeNumber(estimated);
-}
-
-const STALE_STORE_ONLY_CHILD_LINK_MS = 60 * 60 * 1_000;
-const SINGLE_ROW_CONTEXT_CACHE_MAX_ENTRIES = 64;
-
-function isFinitePositiveTimestamp(value: unknown): value is number {
-  return typeof value === "number" && Number.isFinite(value) && value > 0;
-}
-
-function shouldKeepStoreOnlyChildLink(entry: SessionEntry, now: number): boolean {
-  if (isTerminalSessionStatus(entry.status) || isFinitePositiveTimestamp(entry.endedAt)) {
-    const endedAt = isFinitePositiveTimestamp(entry.endedAt) ? entry.endedAt : entry.updatedAt;
-    return (
-      isFinitePositiveTimestamp(endedAt) && now - endedAt <= RECENT_ENDED_SUBAGENT_CHILD_SESSION_MS
-    );
-  }
-  if (entry.status === "running" || isFinitePositiveTimestamp(entry.startedAt)) {
-    return true;
-  }
-  // Store-only child links lack a live subagent registry entry. Keep recent
-  // unknown-state rows visible briefly so reloads do not hide fresh children.
-  return (
-    isFinitePositiveTimestamp(entry.updatedAt) &&
-    now - entry.updatedAt <= STALE_STORE_ONLY_CHILD_LINK_MS
-  );
-}
-
-type SessionListRowContext = {
-  subagentRuns: ReturnType<typeof buildSubagentRunReadIndex>;
-  storeChildSessionsByKey: Map<string, string[]>;
-  selectedModelByOverrideRef: Map<string, ReturnType<typeof resolveSessionModelRef>>;
-  // Per-list memoization for deterministic resolvers that scale linearly with
-  // session count but only depend on (provider, model[, agentId]). Sessions
-  // in a single list typically share a small set of those tuples, so caching
-  // here collapses the work to O(unique tuples) per call.
-  thinkingMetadataByModelRef: Map<
-    string,
-    {
-      levels: ReturnType<typeof listThinkingLevelOptions>;
-      defaultLevel: ReturnType<typeof resolveGatewaySessionThinkingDefault>;
-    }
-  >;
-  displayModelIdentityByKey: Map<string, { provider?: string; model?: string }>;
-  modelCostConfigByModelRef: Map<string, ModelCostConfig | undefined>;
-  userProfileLabelById: Map<string, string | undefined>;
-};
-
-type SessionListRowContextProvider = () => SessionListRowContext;
-
-type SingleRowChildSessionCandidateCacheEntry = {
-  store: Record<string, SessionEntry>;
-  childSessionCandidatesByParentKey: Map<string, string[]>;
-};
-
-type GatewaySessionStoreTarget = {
-  agentId: string;
-  storePath: string;
-  canonicalKey: string;
-  storeKeys: string[];
-};
-
-type GatewaySessionStoreTargetWithStore = GatewaySessionStoreTarget & {
-  store: Record<string, SessionEntry>;
-};
-
-const singleRowChildSessionCandidateCache = new Map<
-  string,
-  SingleRowChildSessionCandidateCacheEntry
->();
-
-function rememberSingleRowChildSessionCandidateCacheEntry(
-  storePath: string,
-  entry: SingleRowChildSessionCandidateCacheEntry,
-) {
-  if (singleRowChildSessionCandidateCache.has(storePath)) {
-    singleRowChildSessionCandidateCache.delete(storePath);
-  }
-  singleRowChildSessionCandidateCache.set(storePath, entry);
-  if (singleRowChildSessionCandidateCache.size <= SINGLE_ROW_CONTEXT_CACHE_MAX_ENTRIES) {
-    return;
-  }
-  const oldestKey = singleRowChildSessionCandidateCache.keys().next().value;
-  if (oldestKey) {
-    singleRowChildSessionCandidateCache.delete(oldestKey);
-  }
-}
-
-function buildStoreChildSessionCandidateIndex(
-  store: Record<string, SessionEntry> | null | undefined,
-): Map<string, string[]> {
-  const childSessionsByKey = new Map<string, string[]>();
-  if (!store) {
-    return childSessionsByKey;
-  }
-  for (const [key, entry] of Object.entries(store)) {
-    if (!entry) {
-      continue;
-    }
-    const parentKeys = [
-      normalizeOptionalString(entry.spawnedBy),
-      normalizeOptionalString(entry.parentSessionKey),
-    ].filter((value): value is string => Boolean(value) && value !== key);
-    for (const parentKey of parentKeys) {
-      addChildSessionKey(childSessionsByKey, parentKey, key);
-    }
-  }
-  return childSessionsByKey;
-}
-
-function getSingleRowChildSessionCandidates(params: {
-  storePath: string;
-  store: Record<string, SessionEntry> | null | undefined;
-}): Map<string, string[]> {
-  if (!params.store) {
-    return new Map();
-  }
-  const cached = singleRowChildSessionCandidateCache.get(params.storePath);
-  if (cached?.store === params.store) {
-    return cached.childSessionCandidatesByParentKey;
-  }
-  const childSessionCandidatesByParentKey = buildStoreChildSessionCandidateIndex(params.store);
-  rememberSingleRowChildSessionCandidateCacheEntry(params.storePath, {
-    store: params.store,
-    childSessionCandidatesByParentKey,
-  });
-  return childSessionCandidatesByParentKey;
-}
-
-function resolveRuntimeChildSessionKeys(
-  controllerSessionKey: string,
-  now = Date.now(),
-  subagentRuns?: SessionListRowContext["subagentRuns"],
-): string[] | undefined {
-  const childSessionKeys = new Set<string>();
-  const controllerKey = controllerSessionKey.trim();
-  const runs = subagentRuns
-    ? (subagentRuns.runsByControllerSessionKey.get(controllerKey) ?? [])
-    : listSubagentRunsForController(controllerSessionKey);
-  for (const entry of runs) {
-    const childSessionKey = normalizeOptionalString(entry.childSessionKey);
-    if (!childSessionKey) {
-      continue;
-    }
-    const latest = subagentRuns
-      ? subagentRuns.getDisplaySubagentRun(childSessionKey)
-      : getSessionDisplaySubagentRunByChildSessionKey(childSessionKey);
-    if (!latest) {
-      continue;
-    }
-    const latestControllerSessionKey =
-      normalizeOptionalString(latest?.controllerSessionKey) ||
-      normalizeOptionalString(latest?.requesterSessionKey);
-    if (latestControllerSessionKey !== controllerSessionKey) {
-      continue;
-    }
-    if (
-      !shouldKeepSubagentRunChildLink(latest, {
-        activeDescendants: subagentRuns
-          ? subagentRuns.countActiveDescendantRuns(childSessionKey)
-          : countActiveDescendantRuns(childSessionKey),
-        now,
-      })
-    ) {
-      continue;
-    }
-    childSessionKeys.add(childSessionKey);
-  }
-  const childSessions = Array.from(childSessionKeys);
-  return childSessions.length > 0 ? childSessions : undefined;
-}
-
-function addChildSessionKey(
-  childSessionsByKey: Map<string, string[]>,
-  parentKey: string,
-  childKey: string,
-) {
-  const current = childSessionsByKey.get(parentKey);
-  if (current) {
-    if (!current.includes(childKey)) {
-      current.push(childKey);
-    }
-    return;
-  }
-  childSessionsByKey.set(parentKey, [childKey]);
-}
-
-function buildStoreChildSessionIndex(
-  store: Record<string, SessionEntry>,
-  now = Date.now(),
-  subagentRuns?: SessionListRowContext["subagentRuns"],
-): Map<string, string[]> {
-  const childSessionsByKey = new Map<string, string[]>();
-  for (const [key, entry] of Object.entries(store)) {
-    if (!entry) {
-      continue;
-    }
-    const parentKeys = [
-      normalizeOptionalString(entry.spawnedBy),
-      normalizeOptionalString(entry.parentSessionKey),
-    ].filter((value): value is string => Boolean(value) && value !== key);
-    if (parentKeys.length === 0) {
-      continue;
-    }
-    const latest = subagentRuns
-      ? subagentRuns.getDisplaySubagentRun(key)
-      : getSessionDisplaySubagentRunByChildSessionKey(key);
-    let latestControllerSessionKey: string | undefined;
-    if (latest) {
-      latestControllerSessionKey =
-        normalizeOptionalString(latest.controllerSessionKey) ||
-        normalizeOptionalString(latest.requesterSessionKey);
-      if (
-        !shouldKeepSubagentRunChildLink(latest, {
-          activeDescendants: subagentRuns
-            ? subagentRuns.countActiveDescendantRuns(key)
-            : countActiveDescendantRuns(key),
-          now,
-        })
-      ) {
-        continue;
-      }
-    } else if (!shouldKeepStoreOnlyChildLink(entry, now)) {
-      continue;
-    }
-    for (const parentKey of parentKeys) {
-      if (latestControllerSessionKey && latestControllerSessionKey !== parentKey) {
-        continue;
-      }
-      addChildSessionKey(childSessionsByKey, parentKey, key);
-    }
-  }
-  return childSessionsByKey;
-}
-
-function resolveStoreChildSessionKeysFromCandidates(params: {
-  store: Record<string, SessionEntry>;
-  key: string;
-  now: number;
-  candidates: ReadonlyMap<string, readonly string[]>;
-}): string[] | undefined {
-  const childSessionKeys: string[] = [];
-  for (const childKey of params.candidates.get(params.key) ?? []) {
-    const entry = params.store[childKey];
-    if (!entry) {
-      continue;
-    }
-    const latest = getSessionDisplaySubagentRunByChildSessionKey(childKey);
-    if (latest) {
-      const latestControllerSessionKey =
-        normalizeOptionalString(latest.controllerSessionKey) ||
-        normalizeOptionalString(latest.requesterSessionKey);
-      if (latestControllerSessionKey !== params.key) {
-        continue;
-      }
-      if (
-        !shouldKeepSubagentRunChildLink(latest, {
-          activeDescendants: countActiveDescendantRuns(childKey),
-          now: params.now,
-        })
-      ) {
-        continue;
-      }
-      childSessionKeys.push(childKey);
-      continue;
-    }
-    if (!shouldKeepStoreOnlyChildLink(entry, params.now)) {
-      continue;
-    }
-    childSessionKeys.push(childKey);
-  }
-  return childSessionKeys.length > 0 ? childSessionKeys : undefined;
-}
-
-function buildSessionListRowContext(params: {
-  store: Record<string, SessionEntry>;
-  now: number;
-}): SessionListRowContext {
-  const subagentRuns = buildSubagentRunReadIndex(params.now);
-  return buildSessionListRowContextFromParts({
-    subagentRuns,
-    storeChildSessionsByKey: buildStoreChildSessionIndex(params.store, params.now, subagentRuns),
-  });
-}
-
-function buildSessionListRowContextFromParts(params: {
-  subagentRuns: ReturnType<typeof buildSubagentRunReadIndex>;
-  storeChildSessionsByKey: Map<string, string[]>;
-}): SessionListRowContext {
-  return {
-    subagentRuns: params.subagentRuns,
-    storeChildSessionsByKey: params.storeChildSessionsByKey,
-    selectedModelByOverrideRef: new Map(),
-    thinkingMetadataByModelRef: new Map(),
-    displayModelIdentityByKey: new Map(),
-    modelCostConfigByModelRef: new Map(),
-    userProfileLabelById: new Map(),
-  };
-}
-
-function buildSessionListRowMetadataContext(params: { now: number }): SessionListRowContext {
-  return buildSessionListRowContextFromParts({
-    subagentRuns: buildSubagentRunReadIndex(params.now),
-    storeChildSessionsByKey: new Map(),
-  });
-}
-
-function buildSingleRowStoreChildSessionsByKey(params: {
-  store: Record<string, SessionEntry>;
-  storePath: string;
-  key: string;
-  now: number;
-}): Map<string, string[]> {
-  const storeChildSessions = resolveStoreChildSessionKeysFromCandidates({
-    store: params.store,
-    key: params.key,
-    now: params.now,
-    candidates: getSingleRowChildSessionCandidates({
-      storePath: params.storePath,
-      store: params.store,
-    }),
-  });
-  return storeChildSessions ? new Map([[params.key, storeChildSessions]]) : new Map();
-}
-
-function createSessionRowModelCacheKey(provider: string | undefined, model: string | undefined) {
-  return `${normalizeLowercaseStringOrEmpty(provider)}\0${normalizeOptionalString(model) ?? ""}`;
-}
-
-function resolveSessionSelectedModelRef(params: {
-  cfg: OpenClawConfig;
-  entry?: SessionEntry;
-  agentId: string;
-  rowContext?: SessionListRowContext;
-  allowPluginNormalization?: boolean;
-}): ReturnType<typeof resolveSessionModelRef> | null {
-  const override = normalizeStoredOverrideModel({
-    providerOverride: params.entry?.providerOverride,
-    modelOverride: params.entry?.modelOverride,
-  });
-  if (!override.modelOverride) {
-    return null;
-  }
-  if (!params.rowContext) {
-    return resolveSessionModelRef(params.cfg, params.entry, params.agentId, {
-      allowPluginNormalization: params.allowPluginNormalization,
-    });
-  }
-  const key = [
-    normalizeAgentId(params.agentId),
-    override.providerOverride ?? "",
-    override.modelOverride,
-  ].join("\0");
-  const cached = params.rowContext.selectedModelByOverrideRef.get(key);
-  if (cached) {
-    return cached;
-  }
-  const selected = resolveSessionModelRef(params.cfg, params.entry, params.agentId, {
-    allowPluginNormalization: params.allowPluginNormalization,
-  });
-  params.rowContext.selectedModelByOverrideRef.set(key, selected);
-  return selected;
-}
-
-function resolveSessionRowThinkingMetadata(params: {
-  cfg: OpenClawConfig;
-  agentId: string;
-  provider: string;
-  model: string;
-  agentRuntime: string;
-  modelCatalog?: ModelCatalogEntry[];
-  rowContext?: SessionListRowContext;
-}): {
-  levels: ReturnType<typeof listThinkingLevelOptions>;
-  defaultLevel: ReturnType<typeof resolveGatewaySessionThinkingDefault>;
-} {
-  if (!params.rowContext) {
-    return {
-      levels: listThinkingLevelOptions(
-        params.provider,
-        params.model,
-        params.modelCatalog,
-        params.agentRuntime,
-      ),
-      defaultLevel: resolveGatewaySessionThinkingDefault({
-        cfg: params.cfg,
-        provider: params.provider,
-        model: params.model,
-        agentId: params.agentId,
-        modelCatalog: params.modelCatalog,
-        agentRuntime: params.agentRuntime,
-      }),
-    };
-  }
-  const key = `${normalizeAgentId(params.agentId)}\0${params.agentRuntime}\0${createSessionRowModelCacheKey(
-    params.provider,
-    params.model,
-  )}`;
-  const cached = params.rowContext.thinkingMetadataByModelRef.get(key);
-  if (cached) {
-    return cached;
-  }
-  const metadata = {
-    levels: listThinkingLevelOptions(
-      params.provider,
-      params.model,
-      params.modelCatalog,
-      params.agentRuntime,
-    ),
-    defaultLevel: resolveGatewaySessionThinkingDefault({
-      cfg: params.cfg,
-      provider: params.provider,
-      model: params.model,
-      agentId: params.agentId,
-      modelCatalog: params.modelCatalog,
-      agentRuntime: params.agentRuntime,
-    }),
-  };
-  params.rowContext.thinkingMetadataByModelRef.set(key, metadata);
-  return metadata;
-}
-
-function mergeChildSessionKeys(
-  runtimeChildSessions: string[] | undefined,
-  storeChildSessions: string[] | undefined,
-): string[] | undefined {
-  if (!runtimeChildSessions?.length) {
-    return storeChildSessions?.length ? storeChildSessions : undefined;
-  }
-  if (!storeChildSessions?.length) {
-    return runtimeChildSessions;
-  }
-  return uniqueStrings([...runtimeChildSessions, ...storeChildSessions]);
-}
-
-function resolveChildSessionKeys(
-  controllerSessionKey: string,
-  store: Record<string, SessionEntry>,
-  now = Date.now(),
-  subagentRuns?: SessionListRowContext["subagentRuns"],
-): string[] | undefined {
-  const runtimeChildSessions = resolveRuntimeChildSessionKeys(
-    controllerSessionKey,
-    now,
-    subagentRuns,
-  );
-  const storeChildSessions = buildStoreChildSessionIndex(store, now, subagentRuns).get(
-    controllerSessionKey,
-  );
-  return mergeChildSessionKeys(runtimeChildSessions, storeChildSessions);
-}
-
-function resolveTranscriptUsageFallback(params: {
-  cfg: OpenClawConfig;
-  key: string;
-  entry?: SessionEntry;
-  storePath: string;
-  fallbackProvider?: string;
-  fallbackModel?: string;
-  maxTranscriptBytes?: number;
-  rowContext?: SessionListRowContext;
-  agentId?: string;
-}): {
-  estimatedCostUsd?: number;
-  totalTokens?: number;
-  totalTokensFresh?: boolean;
-  contextTokens?: number;
-  modelProvider?: string;
-  model?: string;
-} | null {
-  const entry = params.entry;
-  if (!entry?.sessionId) {
-    return null;
-  }
-  const parsed = parseAgentSessionKey(params.key);
-  const agentId = parsed?.agentId
-    ? normalizeAgentId(parsed.agentId)
-    : normalizeAgentId(params.agentId ?? resolveDefaultAgentId(params.cfg));
-  const storePath =
-    resolveConcreteSessionStorePath(params.storePath) ??
-    resolveStorePath(params.cfg.session?.store, { agentId });
-  let snapshot: ReturnType<typeof readScopedRecentSessionUsageFromTranscript>;
-  try {
-    snapshot = readScopedRecentSessionUsageFromTranscript(
-      {
-        agentId,
-        sessionEntry: entry,
-        sessionId: entry.sessionId,
-        sessionKey: params.key,
-        storePath,
-      },
-      typeof params.maxTranscriptBytes === "number" ? params.maxTranscriptBytes : 256 * 1024,
-    );
-  } catch {
-    return null;
-  }
-  if (!snapshot) {
-    return null;
-  }
-  const modelProvider = snapshot.modelProvider ?? params.fallbackProvider;
-  const model = snapshot.model ?? params.fallbackModel;
-  const contextTokens = resolveContextTokensForModel({
-    cfg: params.cfg,
-    provider: modelProvider,
-    model,
-    // Gateway/session listing is read-only; don't start async model discovery.
-    allowAsyncLoad: false,
-  });
-  const estimatedCostUsd = resolveEstimatedSessionCostUsd({
-    cfg: params.cfg,
-    provider: modelProvider,
-    model,
-    explicitCostUsd: snapshot.costUsd,
-    entry: {
-      inputTokens: snapshot.inputTokens,
-      outputTokens: snapshot.outputTokens,
-      cacheRead: snapshot.cacheRead,
-      cacheWrite: snapshot.cacheWrite,
-    },
-    rowContext: params.rowContext,
-  });
-  return {
-    modelProvider,
-    model,
-    totalTokens: resolvePositiveNumber(snapshot.totalTokens),
-    totalTokensFresh: snapshot.totalTokensFresh === true,
-    contextTokens: resolvePositiveNumber(contextTokens),
-    estimatedCostUsd,
-  };
-}
-
-function readAcpMetaForDeletedAgentCheck(params: {
-  cfg: OpenClawConfig;
-  sessionKey: string;
-  entry?: Pick<SessionEntry, "acp" | "lifecycleRevision"> | null;
-  acpMetadataSessionKey?: string | null;
-}) {
-  if (params.entry?.acp) {
-    return params.entry.acp;
-  }
-
-  const acpMetadataSessionKey = normalizeOptionalString(params.acpMetadataSessionKey);
-  const directKeys = new Set<string>();
-  if (acpMetadataSessionKey) {
-    directKeys.add(acpMetadataSessionKey);
-  } else {
-    const acpMeta = readAcpSessionMeta({ sessionKey: params.sessionKey, cfg: params.cfg });
-    if (acpMeta) {
-      return acpMeta;
-    }
-  }
-  directKeys.add(params.sessionKey);
-
-  for (const directKey of directKeys) {
-    const acpMeta = readAcpSessionMetaForEntry({
-      sessionKey: directKey,
-      entry: params.entry ?? undefined,
-    });
-    if (acpMeta) {
-      return acpMeta;
-    }
-  }
-
-  repairAcpSessionMetaKeyForMigration({
-    sessionKey: params.sessionKey,
-    candidateSessionKeys: directKeys,
-    entry: params.entry ?? undefined,
-  });
-  return readAcpSessionMetaForEntry({
-    sessionKey: params.sessionKey,
-    entry: params.entry ?? undefined,
-  });
-}
-
-/**
- * Returns the owning agent id if the session key belongs to an agent that is no
- * longer present in config (deleted). Returns null for non-agent legacy/global
- * keys, confirmed ACP runtime session keys, or when the owning agent still
- * exists (#65524).
- */
-export function resolveDeletedAgentIdFromSessionKey(
-  cfg: OpenClawConfig,
-  sessionKey: string,
-  entry?: SessionEntry | null,
-  options?: { acpMetadataSessionKey?: string | null },
-): string | null {
-  const parsed = parseAgentSessionKey(sessionKey);
-  if (!parsed) {
-    return null;
-  }
-  const agentId = normalizeAgentId(parsed.agentId);
-  if (listAgentIds(cfg).includes(agentId)) {
-    return null;
-  }
-  if (isAcpSessionKey(sessionKey) && !parsed.rest.startsWith("acp:binding:")) {
-    // Free ACP runtime keys use agent:<harnessId>:acp:<uuid>, but key shape is
-    // not proof: ACP bridge sessions can use ACP-shaped keys without SessionAcpMeta.
-    // Configured acp:binding keys stay owner-scoped even when ACP metadata exists.
-    const acpMeta = readAcpMetaForDeletedAgentCheck({
-      cfg,
-      sessionKey,
-      entry,
-      acpMetadataSessionKey: options?.acpMetadataSessionKey,
-    });
-    if (acpMeta) {
-      return null;
-    }
-  }
-  return agentId;
-}
-
-function loadSessionEntryWithMode(
-  sessionKey: string,
-  opts: { agentId?: string; clone?: boolean } | undefined,
-  readOnly: boolean,
-) {
-  const cfg = getRuntimeConfig();
-  const key = normalizeOptionalString(sessionKey) ?? "";
-  const target = resolveGatewaySessionStoreTargetWithStore({
-    cfg,
-    key,
-    ...(opts?.clone === false ? { clone: false } : {}),
-    ...(opts?.agentId ? { agentId: opts.agentId } : {}),
-    ...(readOnly ? { readOnly: true } : {}),
-  });
-  const storePath = target.storePath;
-  const store = target.store;
-  const freshestMatch = resolveFreshestSessionStoreMatchFromStoreKeys(store, target.storeKeys);
-  const legacyKey = freshestMatch?.key !== target.canonicalKey ? freshestMatch?.key : undefined;
-  return {
-    cfg,
-    storePath,
-    store,
-    entry: freshestMatch?.entry,
-    canonicalKey: target.canonicalKey,
-    storeKeys: target.storeKeys,
-    legacyKey,
-  };
-}
-
-export function loadSessionEntry(sessionKey: string, opts?: { agentId?: string; clone?: boolean }) {
-  return loadSessionEntryWithMode(sessionKey, opts, false);
-}
-
-export function loadSessionEntryReadOnly(
-  sessionKey: string,
-  opts?: { agentId?: string; clone?: boolean },
-) {
-  return loadSessionEntryWithMode(sessionKey, opts, true);
-}
-
-/** Returns both the freshest entry and the exact persisted key that owns it. */
-export function resolveFreshestSessionStoreMatchFromStoreKeys(
-  store: Record<string, SessionEntry>,
-  storeKeys: string[],
-): { key: string; entry: SessionEntry } | undefined {
-  let freshest: { key: string; entry: SessionEntry } | undefined;
-  for (const key of storeKeys) {
-    const entry = store[key];
-    if (!entry) {
-      continue;
-    }
-    const match = { key, entry };
-    if (!freshest || (match.entry.updatedAt ?? 0) > (freshest.entry.updatedAt ?? 0)) {
-      freshest = match;
-    }
-  }
-  return freshest;
-}
-
-export function resolveFreshestSessionEntryFromStoreKeys(
-  store: Record<string, SessionEntry>,
-  storeKeys: string[],
-): SessionEntry | undefined {
-  return resolveFreshestSessionStoreMatchFromStoreKeys(store, storeKeys)?.entry;
-}
-
-function findFreshestStoreMatch(
-  store: Record<string, SessionEntry>,
-  ...candidates: string[]
-): { entry: SessionEntry; key: string } | undefined {
-  const matches = new Map<string, { entry: SessionEntry; key: string }>();
-  for (const candidate of candidates) {
-    const trimmed = normalizeOptionalString(candidate) ?? "";
-    if (!trimmed) {
-      continue;
-    }
-    const exact = store[trimmed];
-    if (exact) {
-      matches.set(trimmed, { entry: exact, key: trimmed });
-    }
-  }
-  if (matches.size === 0) {
-    return undefined;
-  }
-  let freshest: { entry: SessionEntry; key: string } | undefined;
-  for (const match of matches.values()) {
-    if (!freshest || (match.entry.updatedAt ?? 0) > (freshest.entry.updatedAt ?? 0)) {
-      freshest = match;
-    }
-  }
-  return freshest;
-}
-
-/**
- * Remove legacy key variants for one canonical session key.
- * Candidates can include aliases (for example, "agent:ops:main" when canonical is "agent:ops:work").
- */
-function pruneLegacyStoreKeys(params: {
-  store: Record<string, unknown>;
-  canonicalKey: string;
-  candidates: Iterable<string>;
-}) {
-  const keysToDelete = new Set<string>();
-  for (const candidate of params.candidates) {
-    const trimmed = normalizeOptionalString(candidate ?? "") ?? "";
-    if (!trimmed) {
-      continue;
-    }
-    if (trimmed !== params.canonicalKey) {
-      keysToDelete.add(trimmed);
-    }
-  }
-  for (const key of keysToDelete) {
-    delete params.store[key];
-  }
-}
-
-export function migrateAndPruneGatewaySessionStoreKey(params: {
-  cfg: OpenClawConfig;
-  key: string;
-  store: Record<string, SessionEntry>;
-  agentId?: string;
-}) {
-  const target = resolveGatewaySessionStoreTarget({
-    cfg: params.cfg,
-    key: params.key,
-    store: params.store,
-    ...(params.agentId ? { agentId: params.agentId } : {}),
-  });
-  const primaryKey = target.canonicalKey;
-  const freshestMatch = resolveFreshestSessionStoreMatchFromStoreKeys(
-    params.store,
-    target.storeKeys,
-  );
-  if (freshestMatch) {
-    const currentPrimary = params.store[primaryKey];
-    if (!currentPrimary || (freshestMatch.entry.updatedAt ?? 0) > (currentPrimary.updatedAt ?? 0)) {
-      params.store[primaryKey] = freshestMatch.entry;
-    }
-  }
-  pruneLegacyStoreKeys({
-    store: params.store,
-    canonicalKey: primaryKey,
-    candidates: target.storeKeys,
-  });
-  return { target, primaryKey, entry: params.store[primaryKey] };
-}
-
-function classifySessionKey(key: string, entry?: SessionEntry): GatewaySessionRow["kind"] {
-  if (key === "global") {
-    return "global";
-  }
-  if (key === "unknown") {
-    return "unknown";
-  }
-  if (entry?.chatType === "group" || entry?.chatType === "channel") {
-    return "group";
-  }
-  if (key.includes(":group:") || key.includes(":channel:")) {
-    return "group";
-  }
-  return "direct";
-}
-
-function parseGroupKey(
-  key: string,
-): { channel?: string; kind?: "group" | "channel"; id?: string } | null {
-  const agentParsed = parseAgentSessionKey(key);
-  const rawKey = agentParsed?.rest ?? key;
-  const parts = rawKey.split(":").filter(Boolean);
-  if (parts.length >= 3) {
-    const [channel, kind, ...rest] = parts;
-    if (kind === "group" || kind === "channel") {
-      const id = rest.join(":");
-      return { channel, kind, id };
-    }
-  }
-  return null;
-}
-
-function isGroupOrChannelDisplaySession(
-  entry: SessionEntry | undefined,
-  parsed: { kind?: "group" | "channel" } | null,
-): boolean {
-  return (
-    entry?.chatType === "group" ||
-    entry?.chatType === "channel" ||
-    parsed?.kind === "group" ||
-    parsed?.kind === "channel"
-  );
-}
-
-function isStorePathTemplate(store?: string): boolean {
-  return typeof store === "string" && store.includes("{agentId}");
-}
-
-function resolveConcreteSessionStorePath(storePath: string | undefined): string | undefined {
-  const trimmed = storePath?.trim();
-  if (!trimmed || trimmed === "(multiple)" || isStorePathTemplate(trimmed)) {
-    return undefined;
-  }
-  return trimmed;
-}
-
-function normalizeFallbackList(values: readonly string[]): string[] {
-  const out: string[] = [];
-  const seen = new Set<string>();
-  for (const value of values) {
-    const trimmed = value.trim();
-    if (!trimmed) {
-      continue;
-    }
-    const key = normalizeLowercaseStringOrEmpty(trimmed);
-    if (seen.has(key)) {
-      continue;
-    }
-    seen.add(key);
-    out.push(trimmed);
-  }
-  return out;
-}
-
-function resolveGatewayAgentModel(
-  cfg: OpenClawConfig,
-  agentId: string,
-): GatewayAgentRow["model"] | undefined {
-  const primary = resolveAgentEffectiveModelPrimary(cfg, agentId)?.trim();
-  const fallbackOverride = resolveAgentModelFallbacksOverride(cfg, agentId);
-  const defaultFallbacks = resolveAgentModelFallbackValues(cfg.agents?.defaults?.model);
-  const fallbacks = normalizeFallbackList(fallbackOverride ?? defaultFallbacks);
-  if (!primary && fallbacks.length === 0) {
-    return undefined;
-  }
-  return {
-    ...(primary ? { primary } : {}),
-    ...(fallbacks.length > 0 ? { fallbacks } : {}),
-  };
-}
-
-export function listAgentsForGateway(
-  cfg: OpenClawConfig,
-  modelCatalog?: ModelCatalogEntry[],
-  options?: {
-    modelCatalogByAgentId?: ReadonlyMap<string, ModelCatalogEntry[]>;
-    includeSystem?: boolean;
-  },
-): {
-  defaultId: string;
-  mainKey: string;
-  scope: SessionScope;
-  agents: GatewayAgentRow[];
-} {
-  const basic = listGatewayAgentsBasic(cfg);
-  const configuredById = new Map<string, { identity?: GatewayAgentRow["identity"] }>();
-  for (const entry of cfg.agents?.list ?? []) {
-    if (!entry?.id) {
-      continue;
-    }
-    const agentId = normalizeAgentId(entry.id);
-    const avatar = normalizeOptionalString(entry.identity?.avatar);
-    const avatarUrl = resolveAgentAvatarUrlFromSource(cfg, agentId, avatar);
-    const identity = entry.identity
-      ? {
-          name: normalizeOptionalString(entry.identity.name),
-          theme: normalizeOptionalString(entry.identity.theme),
-          emoji: normalizeOptionalString(entry.identity.emoji),
-          avatar,
-          avatarUrl,
-        }
-      : undefined;
-    configuredById.set(agentId, { identity });
-  }
-  const roster = options?.includeSystem
-    ? basic.agents
-    : basic.agents.filter((entry) => entry.kind !== "system");
-  const agents = roster.map((entry) => {
-    const { id } = entry;
-    const meta = configuredById.get(id);
-    const model = resolveGatewayAgentModel(cfg, id);
-    const resolvedModel = resolveDefaultModelForAgent({ cfg, agentId: id });
-    const sessionKey = resolveAgentMainSessionKey({ cfg, agentId: id });
-    const agentRuntime = resolveModelAgentRuntimeMetadata({
-      cfg,
-      agentId: id,
-      provider: resolvedModel.provider,
-      model: resolvedModel.model,
-      sessionKey,
-      acpRuntime: false,
-    });
-    const thinkingRuntime = resolveEffectiveAgentRuntime({
-      cfg,
-      provider: resolvedModel.provider,
-      modelId: resolvedModel.model,
-      agentId: id,
-      sessionKey,
-    });
-    const agentModelCatalog = options?.modelCatalogByAgentId?.get(id) ?? modelCatalog;
-    const thinkingLevels = listThinkingLevelOptions(
-      resolvedModel.provider,
-      resolvedModel.model,
-      agentModelCatalog,
-      thinkingRuntime,
-    );
-    const workspace = resolveAgentWorkspaceDir(cfg, id);
-    // Must mirror the sessions.create worktree preflight: subdirectory workspaces inside a
-    // repo are worktree-capable, so the UI toggle and the create path cannot diverge.
-    const workspaceGit = insideGitCheckout(workspace);
-    return Object.assign(
-      {
-        id,
-        ...(options?.includeSystem ? { kind: entry.kind } : {}),
-        name: entry.name,
-        identity: meta?.identity,
-        workspace,
-        workspaceGit,
-        agentRuntime,
-        thinkingLevels,
-        thinkingOptions: thinkingLevels.map((level) => level.label),
-        thinkingDefault: resolveGatewaySessionThinkingDefault({
-          cfg,
-          provider: resolvedModel.provider,
-          model: resolvedModel.model,
-          agentId: id,
-          modelCatalog: agentModelCatalog,
-          agentRuntime: thinkingRuntime,
-        }),
-      },
-      model ? { model } : {},
-    );
-  });
-  return { defaultId: basic.defaultId, mainKey: basic.mainKey, scope: basic.scope, agents };
-}
-
-function buildGatewaySessionStoreScanTargets(params: {
-  cfg: OpenClawConfig;
-  key: string;
-  canonicalKey: string;
-  agentId: string;
-}): string[] {
-  const targets = new Set<string>();
-  if (params.canonicalKey) {
-    targets.add(params.canonicalKey);
-  }
-  if (params.key && params.key !== params.canonicalKey) {
-    targets.add(params.key);
-  }
-  if (params.canonicalKey === "global" || params.canonicalKey === "unknown") {
-    return [...targets];
-  }
-  const agentMainKey = resolveAgentMainSessionKey({ cfg: params.cfg, agentId: params.agentId });
-  if (params.canonicalKey === agentMainKey) {
-    targets.add(`agent:${params.agentId}:main`);
-  }
-  return [...targets];
-}
-
-function resolveGatewaySessionStoreCandidates(
-  cfg: OpenClawConfig,
-  agentId: string,
-): { existing: SessionStoreTarget[]; fallback: SessionStoreTarget } {
-  const storeConfig = cfg.session?.store;
-  const fallback = {
-    agentId,
-    storePath: resolveStorePath(storeConfig, { agentId }),
-  };
-  return {
-    existing: resolveExistingAgentSessionStoreTargetsSync(cfg, agentId),
-    fallback,
-  };
-}
-
-function loadGatewaySessionLookupStore(
-  storePath: string,
-  clone: boolean | undefined,
-  agentId?: string,
-  options: { readOnly?: boolean } = {},
-): Record<string, SessionEntry> {
-  try {
-    const listEntries = options.readOnly
-      ? listAccessorSessionEntriesReadOnly
-      : listAccessorSessionEntries;
-    return Object.fromEntries(
-      listEntries({
-        ...(agentId ? { agentId } : {}),
-        ...(clone === false ? { clone: false } : {}),
-        storePath,
-      }).map(({ sessionKey, entry }) => [sessionKey, entry]),
-    );
-  } catch {
-    return {};
-  }
-}
-
-function resolveGatewaySessionStoreLookup(params: {
-  cfg: OpenClawConfig;
-  key: string;
-  canonicalKey: string;
-  agentId: string;
-  clone?: boolean;
-  initialStore?: Record<string, SessionEntry>;
-  readOnly?: boolean;
-}): {
-  storePath: string;
-  store: Record<string, SessionEntry>;
-  match: { entry: SessionEntry; key: string } | undefined;
-} {
-  const scanTargets = buildGatewaySessionStoreScanTargets(params);
-  const { existing, fallback } = resolveGatewaySessionStoreCandidates(params.cfg, params.agentId);
-  const configured = isConfiguredSessionStoreAgentId(params.cfg, params.agentId);
-  const candidates = configured
-    ? [fallback, ...existing.filter((target) => target.storePath !== fallback.storePath)]
-    : existing;
-  if (candidates.length === 0) {
-    // Discovery is read-only. Only configured agents may cross the fallback edge that creates a
-    // missing SQLite store; retired/manual agents must already have a discovered store.
-    return {
-      storePath: fallback.storePath,
-      store: {},
-      match: undefined,
-    };
-  }
-  const loadStore = (target: SessionStoreTarget) =>
-    loadGatewaySessionLookupStore(target.storePath, params.clone, target.agentId, {
-      readOnly: params.readOnly || !configured,
-    });
-  const firstCandidate = candidates[0] ?? fallback;
-  let selectedStorePath = firstCandidate.storePath;
-  let selectedStore =
-    params.initialStore && firstCandidate.storePath === fallback.storePath
-      ? params.initialStore
-      : loadStore(firstCandidate);
-  let selectedMatch = findFreshestStoreMatch(selectedStore, ...scanTargets);
-  let selectedUpdatedAt = selectedMatch?.entry.updatedAt ?? Number.NEGATIVE_INFINITY;
-
-  for (let index = 1; index < candidates.length; index += 1) {
-    const candidate = candidates[index];
-    if (!candidate) {
-      continue;
-    }
-    const store = loadStore(candidate);
-    const match = findFreshestStoreMatch(store, ...scanTargets);
-    if (!match) {
-      continue;
-    }
-    const updatedAt = match.entry.updatedAt ?? 0;
-    // Mirror combined-store merge behavior so follow-up mutations target the
-    // same backing store that won the listing merge when ids collide.
-    if (!selectedMatch || updatedAt >= selectedUpdatedAt) {
-      selectedStorePath = candidate.storePath;
-      selectedStore = store;
-      selectedMatch = match;
-      selectedUpdatedAt = updatedAt;
-    }
-  }
-
-  return {
-    storePath: selectedStorePath,
-    store: selectedStore,
-    match: selectedMatch,
-  };
-}
-
-function isAgentScopedSentinelSessionKey(canonicalKey: string): boolean {
-  return canonicalKey === "global" || canonicalKey === "unknown";
-}
-
-function resolveExplicitDeletedLegacyMainStoreTarget(params: {
-  cfg: OpenClawConfig;
-  key: string;
-  clone?: boolean;
-  readOnly?: boolean;
-}): GatewaySessionStoreTargetWithStore | null {
-  const parsed = parseAgentSessionKey(params.key);
-  const legacyAgentId = normalizeAgentId(parsed?.agentId);
-  if (
-    !parsed ||
-    legacyAgentId !== DEFAULT_AGENT_ID ||
-    listAgentIds(params.cfg).includes(legacyAgentId)
-  ) {
-    return null;
-  }
-
-  // Only preserve agent:main:* when it is backed by a discovered deleted-main store.
-  // Shared-store legacy aliases should continue remapping to the configured default agent.
-  const canonicalKey = resolveStoredSessionKeyForAgentStore({
-    cfg: params.cfg,
-    agentId: legacyAgentId,
-    sessionKey: params.key,
-  });
-  const agentMainKey = resolveAgentMainSessionKey({ cfg: params.cfg, agentId: legacyAgentId });
-  const legacyAgentMainKey = `agent:${legacyAgentId}:main`;
-  const lookupSeeds = Array.from(
-    new Set([params.key, canonicalKey, agentMainKey, legacyAgentMainKey]),
-  );
-  let best:
-    | {
-        storePath: string;
-        store: Record<string, SessionEntry>;
-        match: { entry: SessionEntry; key: string };
-      }
-    | undefined;
-  for (const target of resolveAllAgentSessionStoreTargetsSync(params.cfg)) {
-    if (target.agentId !== legacyAgentId) {
-      continue;
-    }
-    const store = loadGatewaySessionLookupStore(target.storePath, params.clone, target.agentId, {
-      readOnly: true,
-    });
-    const match = findFreshestStoreMatch(store, ...lookupSeeds);
-    if (!match) {
-      continue;
-    }
-    if (!best || (match.entry.updatedAt ?? 0) >= (best.match.entry.updatedAt ?? 0)) {
-      best = { storePath: target.storePath, store, match };
-    }
-  }
-  if (!best) {
-    return null;
-  }
-
-  const storeKeys = new Set<string>([canonicalKey]);
-  if (params.key !== canonicalKey) {
-    storeKeys.add(params.key);
-  }
-  storeKeys.add(best.match.key);
-  for (const seed of lookupSeeds) {
-    storeKeys.add(seed);
-  }
-  return {
-    agentId: legacyAgentId,
-    storePath: best.storePath,
-    canonicalKey,
-    storeKeys: Array.from(storeKeys),
-    store: best.store,
-  };
-}
-
-export function resolveGatewaySessionStoreTargetWithStore(params: {
-  cfg: OpenClawConfig;
-  key: string;
-  agentId?: string;
-  clone?: boolean;
-  readOnly?: boolean;
-  store?: Record<string, SessionEntry>;
-}): GatewaySessionStoreTargetWithStore {
-  const key = normalizeOptionalString(params.key) ?? "";
-  const explicitDeletedMainTarget = resolveExplicitDeletedLegacyMainStoreTarget({
-    cfg: params.cfg,
-    key,
-    clone: params.clone,
-    readOnly: params.readOnly,
-  });
-  if (explicitDeletedMainTarget) {
-    return explicitDeletedMainTarget;
-  }
-
-  const requestedAgentId = normalizeOptionalString(params.agentId);
-  const canonicalKey = resolveSessionStoreKey({
-    cfg: params.cfg,
-    sessionKey: key,
-    ...(requestedAgentId ? { storeAgentId: requestedAgentId } : {}),
-  });
-  const agentId =
-    requestedAgentId &&
-    (isAgentScopedSentinelSessionKey(canonicalKey) || !parseAgentSessionKey(key))
-      ? normalizeAgentId(requestedAgentId)
-      : resolveSessionStoreAgentId(params.cfg, canonicalKey);
-  if (isIncognitoSessionKey(canonicalKey)) {
-    const storePath = resolveIncognitoOpenClawAgentSqlitePath({ agentId });
-    // Session resolution may receive arbitrary stale keys; only creation/write
-    // owners may materialize the process-lifetime incognito database.
-    const store = loadGatewaySessionLookupStore(storePath, params.clone, agentId, {
-      readOnly: true,
-    });
-    return {
-      agentId,
-      storePath,
-      canonicalKey,
-      storeKeys: [canonicalKey],
-      store,
-    };
-  }
-  const { storePath, store } = resolveGatewaySessionStoreLookup({
-    cfg: params.cfg,
-    key,
-    canonicalKey,
-    agentId,
-    clone: params.clone,
-    readOnly: params.readOnly,
-    initialStore: params.store,
-  });
-
-  if (canonicalKey === "global" || canonicalKey === "unknown") {
-    const storeKeys = key && key !== canonicalKey ? [canonicalKey, key] : [key];
-    return { agentId, storePath, canonicalKey, storeKeys, store };
-  }
-
-  const storeKeys = new Set<string>(
-    buildGatewaySessionStoreScanTargets({ cfg: params.cfg, key, canonicalKey, agentId }),
-  );
-  return {
-    agentId,
-    storePath,
-    canonicalKey,
-    storeKeys: Array.from(storeKeys),
-    store,
-  };
-}
-
-export function resolveGatewaySessionStoreTarget(params: {
-  cfg: OpenClawConfig;
-  key: string;
-  agentId?: string;
-  clone?: boolean;
-  store?: Record<string, SessionEntry>;
-}): GatewaySessionStoreTarget {
-  const { store: _store, ...target } = resolveGatewaySessionStoreTargetWithStore(params);
-  return target;
-}
-
 export { loadCombinedSessionStoreForGateway } from "../config/sessions/combined-store-gateway.js";
-
-function resolveGatewaySessionThinkingLevel(params: {
-  provider: string;
-  model: string;
-  level: NonNullable<ReturnType<typeof normalizeThinkLevel>>;
-  modelCatalog?: ModelCatalogEntry[];
-  agentRuntime: string;
-}) {
-  const catalogEntry = params.modelCatalog
-    ? findModelCatalogEntry(params.modelCatalog, {
-        provider: params.provider,
-        modelId: params.model,
-      })
-    : undefined;
-  // Lightweight sessions.changed projections intentionally omit the catalog.
-  // Runtime/model patches normalize persisted state with authoritative metadata;
-  // projections must not reinterpret an already-validated level without it.
-  if (!catalogEntry) {
-    return params.level;
-  }
-  return resolveSupportedThinkingLevel({
-    provider: params.provider,
-    model: params.model,
-    level: params.level,
-    catalog: params.modelCatalog,
-    agentRuntime: params.agentRuntime,
-  });
-}
-
-function resolveGatewaySessionThinkingDefault(params: {
-  cfg: OpenClawConfig;
-  provider: string;
-  model: string;
-  agentId?: string;
-  modelCatalog?: ModelCatalogEntry[];
-  agentRuntime: string;
-}) {
-  const agentThinkingDefault = params.agentId
-    ? resolveAgentConfig(params.cfg, params.agentId)?.thinkingDefault
-    : undefined;
-  const defaultLevel =
-    agentThinkingDefault ??
-    resolveThinkingDefault({
-      cfg: params.cfg,
-      provider: params.provider,
-      model: params.model,
-      catalog: params.modelCatalog,
-      agentRuntime: params.agentRuntime,
-    });
-  return resolveGatewaySessionThinkingLevel({
-    provider: params.provider,
-    model: params.model,
-    level: defaultLevel,
-    modelCatalog: params.modelCatalog,
-    agentRuntime: params.agentRuntime,
-  });
-}
-
-type GatewaySessionThinkingProjectionParams = {
-  cfg: OpenClawConfig;
-  provider: string;
-  model: string;
-  agentId: string;
-  sessionKey: string;
-  entry?: SessionEntry;
-  modelCatalog?: ModelCatalogEntry[];
-  rowContext?: SessionListRowContext;
-};
-
-function resolveGatewaySessionThinkingProjectionInternal(
-  params: GatewaySessionThinkingProjectionParams,
-) {
-  const acpMeta = readAcpSessionMeta({ sessionKey: params.sessionKey });
-  const configuredAgentRuntime = resolveModelAgentRuntimeMetadata({
-    cfg: params.cfg,
-    agentId: params.agentId,
-    provider: params.provider,
-    model: params.model,
-    sessionKey: params.sessionKey,
-    acpRuntime: acpMeta != null,
-    acpBackend: acpMeta?.backend,
-  });
-  const persistedAgentRuntime = resolveSessionRuntimeOverrideForProvider({
-    provider: params.provider,
-    entry: params.entry,
-    cfg: params.cfg,
-  });
-  const persistedAgentRuntimeSource: "session" | "session-key" =
-    params.entry?.modelSelectionLocked === true ? "session" : "session-key";
-  const agentRuntime =
-    acpMeta || !persistedAgentRuntime
-      ? configuredAgentRuntime
-      : {
-          id: persistedAgentRuntime,
-          source: persistedAgentRuntimeSource,
-        };
-  const thinkingRuntime = acpMeta
-    ? concretizeAgentRuntime(acpMeta.backend ?? agentRuntime.id)
-    : resolveEffectiveAgentRuntime({
-        cfg: params.cfg,
-        provider: params.provider,
-        modelId: params.model,
-        agentId: params.agentId,
-        sessionKey: params.sessionKey,
-        sessionEntry: params.entry,
-      });
-  const metadata = resolveSessionRowThinkingMetadata({
-    cfg: params.cfg,
-    agentId: params.agentId,
-    provider: params.provider,
-    model: params.model,
-    agentRuntime: thinkingRuntime,
-    modelCatalog: params.modelCatalog,
-    rowContext: params.rowContext,
-  });
-  const storedThinkingLevel = normalizeThinkLevel(params.entry?.thinkingLevel);
-  const thinkingLevel = storedThinkingLevel
-    ? resolveGatewaySessionThinkingLevel({
-        provider: params.provider,
-        model: params.model,
-        level: storedThinkingLevel,
-        modelCatalog: params.modelCatalog,
-        agentRuntime: thinkingRuntime,
-      })
-    : undefined;
-  return {
-    agentRuntime,
-    thinkingLevel,
-    effectiveThinkingLevel: thinkingLevel ?? metadata.defaultLevel,
-    thinkingLevels: metadata.levels,
-    thinkingOptions: metadata.levels.map((level) => level.label),
-    thinkingDefault: metadata.defaultLevel,
-  };
-}
-
-/** Resolve the canonical runtime, selected level, and picker metadata for a session. */
-export function resolveGatewaySessionThinkingProjection(params: {
-  cfg: OpenClawConfig;
-  provider: string;
-  model: string;
-  agentId: string;
-  sessionKey: string;
-  entry?: SessionEntry;
-  modelCatalog?: ModelCatalogEntry[];
-}) {
-  return resolveGatewaySessionThinkingProjectionInternal(params);
-}
-
-export function getSessionDefaults(
-  cfg: OpenClawConfig,
-  modelCatalog?: ModelCatalogEntry[],
-  options?: { allowPluginNormalization?: boolean },
-): GatewaySessionsDefaults {
-  const resolved = resolveConfiguredModelRef({
-    cfg,
-    defaultProvider: DEFAULT_PROVIDER,
-    defaultModel: DEFAULT_MODEL,
-    allowPluginNormalization: options?.allowPluginNormalization,
-  });
-  const contextTokens =
-    cfg.agents?.defaults?.contextTokens ??
-    lookupContextTokens(resolved.model, { allowAsyncLoad: false }) ??
-    DEFAULT_CONTEXT_TOKENS;
-  const agentId = normalizeAgentId(resolveDefaultAgentId(cfg));
-  const sessionKey = resolveAgentMainSessionKey({ cfg, agentId });
-  const agentRuntime = resolveModelAgentRuntimeMetadata({
-    cfg,
-    agentId,
-    provider: resolved.provider,
-    model: resolved.model,
-    sessionKey,
-    acpRuntime: false,
-  });
-  const thinkingRuntime = resolveEffectiveAgentRuntime({
-    cfg,
-    provider: resolved.provider,
-    modelId: resolved.model,
-    agentId,
-    sessionKey,
-  });
-  const thinkingLevels = listThinkingLevelOptions(
-    resolved.provider,
-    resolved.model,
-    modelCatalog,
-    thinkingRuntime,
-  );
-  return {
-    modelProvider: resolved.provider ?? null,
-    model: resolved.model ?? null,
-    contextTokens: contextTokens ?? null,
-    agentRuntime,
-    thinkingLevels,
-    thinkingOptions: thinkingLevels.map((level) => level.label),
-    thinkingDefault: resolveGatewaySessionThinkingDefault({
-      cfg,
-      provider: resolved.provider,
-      model: resolved.model,
-      modelCatalog,
-      agentRuntime: thinkingRuntime,
-    }),
-  };
-}
-
-export async function resolveGatewayModelSupportsImages(params: {
-  loadGatewayModelCatalog: (params?: { readOnly?: boolean }) => Promise<ModelCatalogEntry[]>;
-  provider?: string;
-  model?: string;
-}): Promise<boolean> {
-  if (!params.model) {
-    return true;
-  }
-
-  try {
-    const catalog = await params.loadGatewayModelCatalog({ readOnly: false });
-    const modelEntry = findModelCatalogEntry(catalog, {
-      provider: params.provider,
-      modelId: params.model,
-    });
-    const normalizedProvider = normalizeOptionalLowercaseString(
-      params.provider ?? modelEntry?.provider,
-    );
-    const normalizedCandidates = [
-      normalizeLowercaseStringOrEmpty(params.model),
-      normalizeLowercaseStringOrEmpty(modelEntry?.name),
-    ].filter(Boolean);
-    if (modelEntry) {
-      if (modelSupportsInput(modelEntry, "image")) {
-        return true;
-      }
-      // Legacy safety shim for stale persisted Foundry rows that predate
-      // provider-owned capability normalization.
-      if (
-        normalizedProvider === "microsoft-foundry" &&
-        normalizedCandidates.some(
-          (candidate) =>
-            candidate.startsWith("gpt-") ||
-            candidate.startsWith("o1") ||
-            candidate.startsWith("o3") ||
-            candidate.startsWith("o4") ||
-            candidate === "computer-use-preview",
-        )
-      ) {
-        return true;
-      }
-      if (
-        normalizedProvider === "claude-cli" &&
-        normalizedCandidates.some(
-          (candidate) =>
-            candidate === "opus" ||
-            candidate === "sonnet" ||
-            candidate === "haiku" ||
-            candidate.startsWith("claude-"),
-        )
-      ) {
-        return true;
-      }
-      return false;
-    }
-    if (
-      normalizedProvider === "claude-cli" &&
-      normalizedCandidates.some(
-        (candidate) =>
-          candidate === "opus" ||
-          candidate === "sonnet" ||
-          candidate === "haiku" ||
-          candidate.startsWith("claude-"),
-      )
-    ) {
-      return true;
-    }
-    return false;
-  } catch {
-    return false;
-  }
-}
-
-function resolveSessionDisplayModelIdentityRefCached(params: {
-  cfg: OpenClawConfig;
-  agentId: string;
-  provider?: string;
-  model?: string;
-  rowContext?: SessionListRowContext;
-}): { provider?: string; model?: string } {
-  const ctx = params.rowContext;
-  if (!ctx) {
-    return resolveSessionDisplayModelIdentityRef(params);
-  }
-  const key = `${params.agentId}\u0000${createSessionRowModelCacheKey(
-    params.provider,
-    params.model,
-  )}`;
-  const cached = ctx.displayModelIdentityByKey.get(key);
-  if (cached) {
-    return cached;
-  }
-  const value = resolveSessionDisplayModelIdentityRef(params);
-  ctx.displayModelIdentityByKey.set(key, value);
-  return value;
-}
-
-export function resolveSessionDisplayModelIdentityRef(params: {
-  cfg: OpenClawConfig;
-  agentId: string;
-  provider?: string;
-  model?: string;
-}): { provider?: string; model?: string } {
-  const provider = normalizeOptionalString(params.provider);
-  const model = normalizeOptionalString(params.model);
-  if (!provider || !model || !isCliProvider(provider, params.cfg)) {
-    return { provider, model };
-  }
-
-  const defaultRef = resolveDefaultModelForAgent({ cfg: params.cfg, agentId: params.agentId });
-  if (model.includes("/")) {
-    const parsedModel = parseModelRef(model, defaultRef.provider);
-    if (parsedModel && !isCliProvider(parsedModel.provider, params.cfg)) {
-      return parsedModel;
-    }
-  }
-
-  const inferredProvider = inferUniqueProviderFromConfiguredModels({
-    cfg: params.cfg,
-    model,
-  });
-  if (inferredProvider && !isCliProvider(inferredProvider, params.cfg)) {
-    return { provider: inferredProvider, model };
-  }
-
-  const parsedModel = parseModelRef(model, defaultRef.provider);
-  if (parsedModel && !isCliProvider(parsedModel.provider, params.cfg)) {
-    return parsedModel;
-  }
-
-  return {
-    provider: defaultRef.provider || provider,
-    model,
-  };
-}
-
-/** Adds the current human profile label without persisting rename-prone display data. */
-function projectSessionActor(
-  actor: SessionEntry["createdActor"],
-  userProfileLabelById: Map<string, string | undefined> = new Map(),
-): SessionCreatedActor | undefined {
-  if (!actor) {
-    return undefined;
-  }
-  const id = normalizeOptionalString(actor.id);
-  if (actor.type !== "human" || !id) {
-    return { type: actor.type, ...(id ? { id } : {}) };
-  }
-  let label = userProfileLabelById.get(id);
-  if (!userProfileLabelById.has(id)) {
-    try {
-      label = normalizeOptionalString(getUserProfileListItem(id).displayName);
-    } catch {
-      // Human actors can also be channel sender ids; only profile ids resolve here.
-      label = undefined;
-    }
-    userProfileLabelById.set(id, label);
-  }
-  return { type: actor.type, id, ...(label ? { label } : {}) };
-}
-
-export function buildGatewaySessionRow(params: {
-  cfg: OpenClawConfig;
-  storePath: string;
-  store: Record<string, SessionEntry>;
-  key: string;
-  entry?: SessionEntry;
-  modelCatalog?: ModelCatalogEntry[];
-  now?: number;
-  includeDerivedTitles?: boolean;
-  includeLastMessage?: boolean;
-  transcriptUsageMaxBytes?: number;
-  storeChildSessionsByKey?: Map<string, string[]>;
-  rowContext?: SessionListRowContext;
-  agentId?: string;
-  skipTranscriptUsageFallback?: boolean;
-  lightweightListRow?: boolean;
-}): GatewaySessionRow {
-  const { cfg, storePath, store, key, entry } = params;
-  const lightweight = params.lightweightListRow === true;
-  const now = params.now ?? Date.now();
-  const agentStatus = resolveActiveSessionAgentStatus(entry?.agentStatus, now);
-  const observerDigest =
-    entry?.observerDigest &&
-    // Strictly newer: a run end and restart can share a millisecond, and the
-    // prior run's digest must not project onto the replacement run.
-    (entry.startedAt === undefined || entry.observerDigest.updatedAt > entry.startedAt)
-      ? entry.observerDigest
-      : undefined;
-  const updatedAt = entry?.updatedAt ?? null;
-  const parsed = parseGroupKey(key);
-  const channel = entry?.channel ?? parsed?.channel;
-  const subject = entry?.subject;
-  const groupChannel = entry?.groupChannel;
-  const space = entry?.space;
-  const id = parsed?.id;
-  const origin = entry?.origin;
-  const originLabel = origin?.label;
-  const parsedAgent = parseAgentSessionKey(key);
-  const isDashboardSession = parsedAgent?.rest.startsWith("dashboard:") === true;
-  const isGroupSession = isGroupOrChannelDisplaySession(entry, parsed);
-  // A user-assigned label is an explicit rename; it must win over stored
-  // channel-derived display names or renames silently vanish on refresh.
-  // Group sessions prefer the human chat title (subject/#channel) over the
-  // stored compact token displayName (e.g. "slack:g-general").
-  const displayName =
-    entry?.label ??
-    (isGroupSession ? buildGroupDisplayTitle({ subject, groupChannel, space }) : undefined) ??
-    entry?.displayName ??
-    (isGroupSession && channel
-      ? buildGroupDisplayName({
-          provider: channel,
-          subject,
-          groupChannel,
-          space,
-          id,
-          key,
-        })
-      : undefined) ??
-    // Dashboard origin labels identify the authenticated sender. Using them as
-    // titles leaks account names into the sidebar while the generated title is pending.
-    (isDashboardSession ? undefined : originLabel);
-  const deliveryFields = normalizeSessionDeliveryFields(entry);
-  const sessionAgentId = normalizeAgentId(
-    parsedAgent?.agentId ?? params.agentId ?? resolveDefaultAgentId(cfg),
-  );
-  const skipTranscriptUsage = params.skipTranscriptUsageFallback === true;
-  const rowContext = params.rowContext;
-  const subagentRun = rowContext
-    ? rowContext.subagentRuns.getDisplaySubagentRun(key)
-    : getSessionDisplaySubagentRunByChildSessionKey(key);
-  const subagentOwner =
-    normalizeOptionalString(subagentRun?.controllerSessionKey) ||
-    normalizeOptionalString(subagentRun?.requesterSessionKey);
-  const liveSubagentRunActive = isSubagentRunLive(subagentRun);
-  const persistedSessionStatus = entry?.status;
-  const persistedSessionEndedAt = entry?.endedAt;
-  const persistedSessionStartedAt = entry?.startedAt;
-  const persistedSessionRuntimeMs = entry?.runtimeMs;
-  const subagentRunState = subagentRun
-    ? liveSubagentRunActive
-      ? "active"
-      : typeof subagentRun.endedAt === "number" ||
-          persistedSessionStatus === "done" ||
-          persistedSessionStatus === "failed" ||
-          persistedSessionStatus === "killed" ||
-          persistedSessionStatus === "timeout" ||
-          typeof persistedSessionEndedAt === "number"
-        ? "historical"
-        : "interrupted"
-    : undefined;
-  const subagentStatus = subagentRun
-    ? liveSubagentRunActive
-      ? resolveSubagentSessionStatus(subagentRun)
-      : persistedSessionStatus === "running"
-        ? undefined
-        : (persistedSessionStatus ??
-          (typeof subagentRun.endedAt === "number"
-            ? resolveSubagentSessionStatus(subagentRun)
-            : undefined))
-    : undefined;
-  const subagentStartedAt = subagentRun
-    ? liveSubagentRunActive
-      ? getSubagentSessionStartedAt(subagentRun)
-      : (persistedSessionStartedAt ?? getSubagentSessionStartedAt(subagentRun))
-    : undefined;
-  const subagentEndedAt = subagentRun
-    ? liveSubagentRunActive
-      ? subagentRun.endedAt
-      : (persistedSessionEndedAt ?? subagentRun.endedAt)
-    : undefined;
-  const subagentRuntimeMs = subagentRun
-    ? liveSubagentRunActive
-      ? resolveSessionRuntimeMs(subagentRun, now)
-      : (persistedSessionRuntimeMs ??
-        (typeof subagentRun.endedAt === "number"
-          ? resolveSessionRuntimeMs(subagentRun, now)
-          : undefined))
-    : undefined;
-  const selectedModel = resolveSessionSelectedModelRef({
-    cfg,
-    entry,
-    agentId: sessionAgentId,
-    rowContext,
-    allowPluginNormalization: !lightweight,
-  });
-  const resolvedModel = resolveSessionModelIdentityRef(
-    cfg,
-    entry,
-    sessionAgentId,
-    subagentRun?.model,
-    { allowPluginNormalization: !lightweight },
-  );
-  const runtimeModelPresent =
-    Boolean(entry?.model?.trim()) || Boolean(entry?.modelProvider?.trim());
-  const freshSessionTotalTokens = resolveNonNegativeNumber(resolveFreshSessionTotalTokens(entry));
-  const needsTranscriptTotalTokens = freshSessionTotalTokens === undefined;
-  const needsTranscriptContextTokens = resolvePositiveNumber(entry?.contextTokens) === undefined;
-  const needsTranscriptEstimatedCostUsd =
-    !skipTranscriptUsage &&
-    resolveEstimatedSessionCostUsd({
-      cfg,
-      provider: resolvedModel.provider,
-      model: resolvedModel.model ?? DEFAULT_MODEL,
-      entry,
-      rowContext,
-    }) === undefined;
-  const transcriptUsage =
-    !skipTranscriptUsage &&
-    (needsTranscriptTotalTokens || needsTranscriptContextTokens || needsTranscriptEstimatedCostUsd)
-      ? resolveTranscriptUsageFallback({
-          cfg,
-          key,
-          entry,
-          storePath,
-          fallbackProvider: resolvedModel.provider,
-          fallbackModel: resolvedModel.model ?? DEFAULT_MODEL,
-          maxTranscriptBytes: params.transcriptUsageMaxBytes,
-          rowContext: params.rowContext,
-          agentId: sessionAgentId,
-        })
-      : null;
-  const preferLiveSubagentModelIdentity =
-    Boolean(subagentRun?.model?.trim()) && subagentStatus === "running";
-  const shouldUseTranscriptModelIdentity =
-    runtimeModelPresent &&
-    !preferLiveSubagentModelIdentity &&
-    (needsTranscriptTotalTokens || needsTranscriptContextTokens);
-  const resolvedModelIdentity = {
-    provider: resolvedModel.provider,
-    model: resolvedModel.model ?? DEFAULT_MODEL,
-  };
-  const modelIdentity = shouldUseTranscriptModelIdentity
-    ? {
-        provider: transcriptUsage?.modelProvider ?? resolvedModelIdentity.provider,
-        model: transcriptUsage?.model ?? resolvedModelIdentity.model,
-      }
-    : resolvedModelIdentity;
-  const { provider: modelProvider, model } = modelIdentity;
-  const totalTokens =
-    freshSessionTotalTokens ?? resolveNonNegativeNumber(transcriptUsage?.totalTokens);
-  const totalTokensFresh =
-    freshSessionTotalTokens !== undefined ||
-    (typeof totalTokens === "number" && Number.isFinite(totalTokens) && totalTokens > 0)
-      ? true
-      : transcriptUsage?.totalTokensFresh === true;
-  const goal = entry?.goal
-    ? resolveSessionGoalDisplayState(
-        {
-          goal: entry.goal,
-          totalTokens,
-          totalTokensFresh,
-        },
-        now,
-        // Session listing is read-only; stale goal baselines are adopted only
-        // by goal commands/tools that can persist the first fresh snapshot.
-        { adoptFreshBaseline: false },
-      )
-    : undefined;
-  const childSessions = params.storeChildSessionsByKey
-    ? mergeChildSessionKeys(
-        resolveRuntimeChildSessionKeys(key, now, rowContext?.subagentRuns),
-        params.storeChildSessionsByKey.get(key),
-      )
-    : resolveChildSessionKeys(key, store, now, rowContext?.subagentRuns);
-  const compactionCheckpoints = resolveProjectableCompactionCheckpoints(entry);
-  const compactionCheckpointCount = Array.isArray(entry?.compactionCheckpoints)
-    ? compactionCheckpoints.length
-    : undefined;
-  const latestCompactionCheckpoint = buildCompactionCheckpointPreview(
-    resolveLatestCompactionCheckpoint(compactionCheckpoints),
-  );
-  const selectedOrRuntimeModelProvider = selectedModel?.provider ?? modelProvider;
-  const selectedOrRuntimeModel = selectedModel?.model ?? model;
-  const rowModelIdentity = lightweight
-    ? { provider: selectedOrRuntimeModelProvider, model: selectedOrRuntimeModel }
-    : resolveSessionDisplayModelIdentityRefCached({
-        cfg,
-        agentId: sessionAgentId,
-        provider: selectedOrRuntimeModelProvider,
-        model: selectedOrRuntimeModel,
-        rowContext: params.rowContext,
-      });
-  const rowModelProvider = rowModelIdentity.provider;
-  const rowModel = rowModelIdentity.model;
-  const acpSessionKey = resolveStoredSessionKeyForAgentStore({
-    cfg,
-    agentId: sessionAgentId,
-    sessionKey: key,
-  });
-  const estimatedCostUsd = lightweight
-    ? resolveNonNegativeNumber(entry?.estimatedCostUsd)
-    : (resolveEstimatedSessionCostUsd({
-        cfg,
-        provider: rowModelProvider,
-        model: rowModel,
-        entry,
-        rowContext: params.rowContext,
-      }) ?? resolveNonNegativeNumber(transcriptUsage?.estimatedCostUsd));
-  const contextTokens = lightweight
-    ? (resolvePositiveNumber(entry?.contextTokens) ??
-      resolvePositiveNumber(
-        resolveContextTokensForModel({
-          cfg,
-          provider: rowModelProvider,
-          model: rowModel,
-          allowAsyncLoad: false,
-        }),
-      ))
-    : (resolvePositiveNumber(entry?.contextTokens) ??
-      resolvePositiveNumber(transcriptUsage?.contextTokens) ??
-      resolvePositiveNumber(
-        resolveContextTokensForModel({
-          cfg,
-          provider: rowModelProvider,
-          model: rowModel,
-          allowAsyncLoad: false,
-        }),
-      ));
-
-  let derivedTitle: string | undefined;
-  let lastMessagePreview: string | undefined;
-  if (entry?.sessionId && (params.includeDerivedTitles || params.includeLastMessage)) {
-    const fields = readScopedSessionTitleFieldsFromTranscript({
-      agentId: sessionAgentId,
-      sessionEntry: entry,
-      sessionId: entry.sessionId,
-      sessionKey: key,
-      storePath,
-    });
-    if (params.includeDerivedTitles) {
-      derivedTitle = deriveSessionTitle(entry, fields.firstUserMessage);
-    }
-    if (params.includeLastMessage && fields.lastMessagePreview) {
-      lastMessagePreview = fields.lastMessagePreview;
-    }
-  }
-
-  const thinkingProvider = rowModelProvider ?? DEFAULT_PROVIDER;
-  const thinkingModel = rowModel ?? DEFAULT_MODEL;
-  const thinkingProjection = resolveGatewaySessionThinkingProjectionInternal({
-    cfg,
-    agentId: sessionAgentId,
-    provider: thinkingProvider,
-    model: thinkingModel,
-    sessionKey: acpSessionKey,
-    entry,
-    modelCatalog: params.modelCatalog,
-    rowContext,
-  });
-  const fastModeState = resolveFastModeState({
-    cfg,
-    provider: selectedOrRuntimeModelProvider ?? DEFAULT_PROVIDER,
-    model: selectedOrRuntimeModel ?? DEFAULT_MODEL,
-    agentId: sessionAgentId,
-    sessionEntry:
-      entry?.fastMode !== undefined
-        ? {
-            fastMode: entry.fastMode,
-          }
-        : undefined,
-  });
-  const pluginExtensions =
-    !lightweight && entry ? projectPluginSessionExtensionsSync({ sessionKey: key, entry }) : [];
-
-  return {
-    key,
-    visibility: entry ? (entry.visibility ?? "shared") : undefined,
-    incognito: entry?.incognito,
-    spawnedBy: subagentOwner || entry?.spawnedBy,
-    // The live registry controller takes precedence over the persisted spawner.
-    controlOwnerSessionKey: subagentOwner || entry?.spawnedBy,
-    swarmGroupId: entry?.swarmGroupId,
-    spawnedWorkspaceDir: entry?.spawnedWorkspaceDir,
-    spawnedCwd: entry?.spawnedCwd,
-    worktree: entry?.worktree,
-    execNode: entry?.execNode,
-    execCwd: entry?.execCwd,
-    forkedFromParent: sessionEntryForkedFromParent(entry) ? true : undefined,
-    spawnDepth: entry?.spawnDepth,
-    subagentRole: entry?.subagentRole,
-    subagentControlScope: entry?.subagentControlScope,
-    createdVia: entry?.createdVia,
-    createdActor: projectSessionActor(entry?.createdActor, rowContext?.userProfileLabelById),
-    createdAt: entry?.createdAt,
-    forkSource: entry?.forkSource,
-    previousSessionId: entry?.previousSessionId,
-    kind: classifySessionKey(key, entry),
-    label: entry?.label,
-    category: entry?.category,
-    displayName,
-    derivedTitle,
-    lastMessagePreview,
-    channel,
-    subject,
-    groupChannel,
-    space,
-    chatType: entry?.chatType,
-    origin,
-    updatedAt,
-    archived: entry?.archivedAt !== undefined,
-    archivedAt: entry?.archivedAt,
-    archivedBy: projectSessionActor(entry?.archivedBy, rowContext?.userProfileLabelById),
-    pinned: entry?.pinnedAt !== undefined,
-    pinnedAt: entry?.pinnedAt,
-    icon: entry?.icon,
-    unread: deriveSessionUnread(entry),
-    lastReadAt: entry?.lastReadAt,
-    agentStatus,
-    observerDigest: observerDigest
-      ? {
-          runId: observerDigest.runId,
-          headline: observerDigest.headline,
-          health: observerDigest.health,
-          updatedAt: observerDigest.updatedAt,
-          revision: observerDigest.revision,
-        }
-      : undefined,
-    lastInteractionAt: entry?.lastInteractionAt,
-    lastActivityAt: entry?.lastActivityAt,
-    sessionId: entry?.sessionId,
-    systemSent: entry?.systemSent,
-    abortedLastRun: entry?.abortedLastRun,
-    thinkingLevel: thinkingProjection.thinkingLevel,
-    thinkingLevels: thinkingProjection.thinkingLevels,
-    thinkingOptions: thinkingProjection.thinkingOptions,
-    thinkingDefault: thinkingProjection.thinkingDefault,
-    fastMode: entry?.fastMode,
-    effectiveFastMode: fastModeState.mode,
-    effectiveFastModeSource: fastModeState.source,
-    fastAutoOnSeconds: fastModeState.fastAutoOnSeconds,
-    verboseLevel: entry?.verboseLevel,
-    traceLevel: entry?.traceLevel,
-    reasoningLevel: entry?.reasoningLevel,
-    elevatedLevel: entry?.elevatedLevel,
-    sendPolicy: entry?.sendPolicy,
-    inputTokens: entry?.inputTokens,
-    outputTokens: entry?.outputTokens,
-    totalTokens,
-    totalTokensFresh,
-    goal,
-    estimatedCostUsd,
-    status: subagentRun ? subagentStatus : entry?.status,
-    lastRunError: entry?.lastRunError,
-    hasAutomation: sessionHasAutomation(key, cfg) ? true : undefined,
-    subagentRunState,
-    hasActiveSubagentRun: subagentRun ? liveSubagentRunActive : undefined,
-    startedAt: subagentRun ? subagentStartedAt : entry?.startedAt,
-    endedAt: subagentRun ? subagentEndedAt : entry?.endedAt,
-    runtimeMs: subagentRun ? subagentRuntimeMs : entry?.runtimeMs,
-    // Navigation lineage is persisted; runtime control is exposed separately above.
-    parentSessionKey: entry?.parentSessionKey,
-    childSessions,
-    responseUsage: entry?.responseUsage,
-    effectiveResponseUsage: resolveEffectiveResponseUsage(
-      entry?.responseUsage,
-      cfg.messages?.responseUsage,
-      channel,
-    ),
-    queueMode: entry?.queueMode,
-    effectiveQueueMode: resolveQueueSettings({
-      cfg,
-      channel: INTERNAL_MESSAGE_CHANNEL,
-      sessionEntry: entry,
-    }).mode,
-    modelProvider: rowModelProvider,
-    model: rowModel,
-    modelSelectionLocked: entry?.modelSelectionLocked,
-    agentRuntime: thinkingProjection.agentRuntime,
-    contextTokens,
-    contextBudgetStatus: entry?.contextBudgetStatus,
-    deliveryContext: deliveryFields.deliveryContext,
-    lastChannel: deliveryFields.lastChannel ?? entry?.lastChannel,
-    lastTo: deliveryFields.lastTo ?? entry?.lastTo,
-    lastAccountId: deliveryFields.lastAccountId ?? entry?.lastAccountId,
-    lastThreadId: deliveryFields.lastThreadId ?? entry?.lastThreadId,
-    compactionCheckpointCount,
-    latestCompactionCheckpoint,
-    pluginExtensions: pluginExtensions.length > 0 ? pluginExtensions : undefined,
-  };
-}
-
-function resolveSessionListSearchDisplayName(
-  key: string,
-  entry?: SessionEntry,
-): string | undefined {
-  if (entry?.displayName) {
-    return entry.displayName;
-  }
-  const parsed = parseGroupKey(key);
-  const channel = entry?.channel ?? parsed?.channel;
-  if (isGroupOrChannelDisplaySession(entry, parsed) && channel) {
-    return buildGroupDisplayName({
-      provider: channel,
-      subject: entry?.subject,
-      groupChannel: entry?.groupChannel,
-      space: entry?.space,
-      id: parsed?.id,
-      key,
-    });
-  }
-  return entry?.label ?? entry?.origin?.label;
-}
-
-function addSessionListSearchModelFields(
-  fields: Array<string | undefined>,
-  identity: { provider?: string; model?: string },
-) {
-  const provider = normalizeOptionalString(identity.provider);
-  const model = normalizeOptionalString(identity.model);
-  fields.push(provider, model);
-  if (provider && model) {
-    fields.push(`${provider}/${model}`);
-  }
-}
-
-function matchesSessionListSearch(fields: Array<string | undefined>, search: string): boolean {
-  return fields.some(
-    (field) => typeof field === "string" && normalizeLowercaseStringOrEmpty(field).includes(search),
-  );
-}
-
-function appendStoredSessionModelSearchFields(
-  fields: Array<string | undefined>,
-  entry?: SessionEntry,
-) {
-  const provider = normalizeOptionalString(entry?.modelProvider);
-  const model = normalizeOptionalString(entry?.model);
-  fields.push(provider, model);
-  if (provider && model) {
-    fields.push(`${provider}/${model}`);
-  }
-}
-
-function shouldResolveDerivedSessionModelSearchFields(search: string): boolean {
-  // Agent session-key searches are already covered by cheap key fields; do not
-  // hydrate model metadata for every non-matching row on hot TUI lookups.
-  return !search.startsWith("agent:");
-}
-
-function resolveSessionListRowContext(params: {
-  rowContext?: SessionListRowContext;
-  getRowContext?: SessionListRowContextProvider;
-}): SessionListRowContext | undefined {
-  return params.rowContext ?? params.getRowContext?.();
-}
-
-function resolveSessionListSearchModelFields(params: {
-  cfg: OpenClawConfig;
-  key: string;
-  entry?: SessionEntry;
-  rowContext?: SessionListRowContext;
-}): Array<string | undefined> {
-  const parsedAgent = parseAgentSessionKey(params.key);
-  const agentId = normalizeAgentId(parsedAgent?.agentId ?? resolveDefaultAgentId(params.cfg));
-  const subagentRun = params.rowContext
-    ? params.rowContext.subagentRuns.getDisplaySubagentRun(params.key)
-    : getSessionDisplaySubagentRunByChildSessionKey(params.key);
-  const selectedModel = resolveSessionSelectedModelRef({
-    cfg: params.cfg,
-    entry: params.entry,
-    agentId,
-    rowContext: params.rowContext,
-    allowPluginNormalization: false,
-  });
-  const resolvedModel = resolveSessionModelIdentityRef(
-    params.cfg,
-    params.entry,
-    agentId,
-    subagentRun?.model,
-    { allowPluginNormalization: false },
-  );
-  const modelIdentity = {
-    provider: resolvedModel.provider,
-    model: resolvedModel.model ?? DEFAULT_MODEL,
-  };
-  const selectedOrRuntimeModelProvider = selectedModel?.provider ?? modelIdentity.provider;
-  const selectedOrRuntimeModel = selectedModel?.model ?? modelIdentity.model;
-  const displayModelIdentity = resolveSessionDisplayModelIdentityRefCached({
-    cfg: params.cfg,
-    agentId,
-    provider: selectedOrRuntimeModelProvider,
-    model: selectedOrRuntimeModel,
-    rowContext: params.rowContext,
-  });
-  const fields: Array<string | undefined> = [];
-  addSessionListSearchModelFields(fields, {
-    provider: params.entry?.modelProvider,
-    model: params.entry?.model,
-  });
-  addSessionListSearchModelFields(fields, resolvedModel);
-  if (selectedModel) {
-    addSessionListSearchModelFields(fields, selectedModel);
-  }
-  addSessionListSearchModelFields(fields, displayModelIdentity);
-  return fields;
-}
-
-export function loadGatewaySessionRow(
-  sessionKey: string,
-  options?: {
-    agentId?: string;
-    includeDerivedTitles?: boolean;
-    includeLastMessage?: boolean;
-    now?: number;
-    transcriptUsageMaxBytes?: number;
-  },
-): GatewaySessionRow | null {
-  const now = options?.now ?? Date.now();
-  const { cfg, storePath, store, entry, canonicalKey } = loadSessionEntryReadOnly(sessionKey, {
-    clone: false,
-    ...(options?.agentId ? { agentId: options.agentId } : {}),
-  });
-  if (!entry) {
-    return null;
-  }
-  const storeChildSessionsByKey = buildSingleRowStoreChildSessionsByKey({
-    storePath,
-    store,
-    key: canonicalKey,
-    now,
-  });
-  return buildGatewaySessionRow({
-    cfg,
-    storePath,
-    store,
-    key: canonicalKey,
-    entry,
-    now,
-    includeDerivedTitles: options?.includeDerivedTitles,
-    includeLastMessage: options?.includeLastMessage,
-    transcriptUsageMaxBytes: options?.transcriptUsageMaxBytes,
-    storeChildSessionsByKey,
-    ...(options?.agentId ? { agentId: options.agentId } : {}),
-  });
-}
-
-export function buildGatewaySessionInfo(params: {
-  cfg: OpenClawConfig;
-  storePath: string;
-  store: Record<string, SessionEntry>;
-  key: string;
-  entry?: SessionEntry;
-  agentId?: string;
-  now?: number;
-  modelCatalog?: ModelCatalogEntry[];
-}): GatewaySessionRow {
-  const now = params.now ?? Date.now();
-  const storeChildSessionsByKey = buildSingleRowStoreChildSessionsByKey({
-    storePath: params.storePath,
-    store: params.store,
-    key: params.key,
-    now,
-  });
-  return buildGatewaySessionRow({
-    cfg: params.cfg,
-    storePath: params.storePath,
-    store: params.store,
-    key: params.key,
-    entry: params.entry,
-    agentId: params.agentId,
-    modelCatalog: params.modelCatalog,
-    now,
-    storeChildSessionsByKey,
-    skipTranscriptUsageFallback: true,
-    lightweightListRow: true,
-  });
-}
-
-/**
- * Number of session rows to build per batch before yielding to the event loop.
- * Keeps the main thread responsive during large session list operations while
- * avoiding excessive yielding overhead for small stores.
- */
-const SESSIONS_LIST_YIELD_BATCH_SIZE = 10;
-const SESSIONS_LIST_DEFAULT_LIMIT = 100;
-
-type SessionEntrySelection = {
-  entries: SessionEntryPair[];
-  creatorEntries: SessionEntryPair[];
-  totalCount: number;
-  limitApplied?: number;
-  offset: number;
-  nextOffset: number | null;
-  hasMore: boolean;
-};
-
-function resolveSessionsListLimit(
-  opts: SessionsListParams,
-  defaultLimit?: number,
-): number | undefined {
-  if (typeof opts.limit !== "number" || !Number.isFinite(opts.limit)) {
-    return defaultLimit;
-  }
-  return Math.max(1, Math.floor(opts.limit));
-}
-
-function resolveSessionsListOffset(opts: SessionsListParams): number {
-  if (typeof opts.offset !== "number" || !Number.isFinite(opts.offset)) {
-    return 0;
-  }
-  return Math.max(0, Math.floor(opts.offset));
-}
-
-function resolveSessionsListWindowLimit(limit: number | undefined, offset: number) {
-  if (limit === undefined) {
-    return undefined;
-  }
-  const windowLimit = offset + limit;
-  return Number.isFinite(windowLimit) ? Math.min(windowLimit, Number.MAX_SAFE_INTEGER) : undefined;
-}
-
-function filterSessionEntries(params: {
-  cfg: OpenClawConfig;
-  store: Record<string, SessionEntry>;
-  opts: SessionsListParams;
-  now: number;
-  rowContext?: SessionListRowContext;
-  getRowContext?: SessionListRowContextProvider;
-}): SessionEntryPair[] {
-  const { cfg, store, opts, now } = params;
-  const includeGlobal = opts.includeGlobal === true;
-  const includeUnknown = opts.includeUnknown === true;
-  const spawnedBy = typeof opts.spawnedBy === "string" ? opts.spawnedBy : "";
-  const label = normalizeOptionalString(opts.label) ?? "";
-  const agentId = typeof opts.agentId === "string" ? normalizeAgentId(opts.agentId) : "";
-  const search = normalizeLowercaseStringOrEmpty(opts.search);
-  const activeMinutes =
-    typeof opts.activeMinutes === "number" && Number.isFinite(opts.activeMinutes)
-      ? Math.max(1, Math.floor(opts.activeMinutes))
-      : undefined;
-
-  let entries = Object.entries(store)
-    .filter(([key]) => {
-      if (isCronRunSessionKey(key)) {
-        return false;
-      }
-      if (!includeGlobal && key === "global") {
-        return false;
-      }
-      if (!includeUnknown && key === "unknown") {
-        return false;
-      }
-      if (agentId) {
-        if (key === "global") {
-          return includeGlobal;
-        }
-        if (key === "unknown") {
-          return false;
-        }
-        const parsed = parseAgentSessionKey(key);
-        if (!parsed) {
-          return false;
-        }
-        return normalizeAgentId(parsed.agentId) === agentId;
-      }
-      return true;
-    })
-    .filter(([key, entry]) => {
-      if (isPhantomAgentStoreListEntry(key, entry)) {
-        return false;
-      }
-      if (!spawnedBy) {
-        return true;
-      }
-      if (key === "unknown" || key === "global") {
-        return false;
-      }
-      const filterRowContext = resolveSessionListRowContext(params);
-      const latest = filterRowContext
-        ? filterRowContext.subagentRuns.getDisplaySubagentRun(key)
-        : getSessionDisplaySubagentRunByChildSessionKey(key);
-      if (latest) {
-        const latestControllerSessionKey =
-          normalizeOptionalString(latest.controllerSessionKey) ||
-          normalizeOptionalString(latest.requesterSessionKey);
-        return (
-          latestControllerSessionKey === spawnedBy &&
-          shouldKeepSubagentRunChildLink(latest, {
-            activeDescendants: filterRowContext
-              ? filterRowContext.subagentRuns.countActiveDescendantRuns(key)
-              : countActiveDescendantRuns(key),
-            now,
-          })
-        );
-      }
-      return (
-        shouldKeepStoreOnlyChildLink(entry, now) &&
-        (entry?.spawnedBy === spawnedBy || entry?.parentSessionKey === spawnedBy)
-      );
-    })
-    .filter(([, entry]) => {
-      if (opts.archived === "all") {
-        return true;
-      }
-      const archived = entry?.archivedAt !== undefined;
-      return opts.archived === true ? archived : !archived;
-    })
-    .filter(([, entry]) => {
-      if (opts.requireLastInteraction !== true) {
-        return true;
-      }
-      return (
-        isFinitePositiveTimestamp(entry?.lastInteractionAt) &&
-        !normalizeOptionalString(entry?.heartbeatIsolatedBaseSessionKey)
-      );
-    })
-    .filter(([, entry]) => {
-      if (!label) {
-        return true;
-      }
-      return entry?.label === label;
-    });
-
-  if (search) {
-    entries = entries.filter(([key, entry]) => {
-      const cheapFields = [
-        resolveSessionListSearchDisplayName(key, entry),
-        entry?.label,
-        entry?.subject,
-        entry?.sessionId,
-        key,
-      ];
-      appendStoredSessionModelSearchFields(cheapFields, entry);
-      if (matchesSessionListSearch(cheapFields, search)) {
-        return true;
-      }
-      if (!shouldResolveDerivedSessionModelSearchFields(search)) {
-        return false;
-      }
-      const searchRowContext = resolveSessionListRowContext(params);
-      return matchesSessionListSearch(
-        resolveSessionListSearchModelFields({
-          cfg,
-          key,
-          entry,
-          rowContext: searchRowContext,
-        }),
-        search,
-      );
-    });
-  }
-
-  if (activeMinutes !== undefined) {
-    const cutoff = now - activeMinutes * 60_000;
-    entries = entries.filter(([, entry]) => (entry?.updatedAt ?? 0) >= cutoff);
-  }
-
-  return entries;
-}
-
-function isPhantomAgentStoreListEntry(key: string, entry: SessionEntry | undefined): boolean {
-  const parsed = parseAgentSessionKey(key);
-  return (
-    parsed?.rest === "sessions" &&
-    !normalizeOptionalString(entry?.sessionId) &&
-    entry?.updatedAt == null
-  );
-}
-
-function selectSessionEntries(params: {
-  cfg: OpenClawConfig;
-  store: Record<string, SessionEntry>;
-  opts: SessionsListParams;
-  now: number;
-  rowContext?: SessionListRowContext;
-  getRowContext?: SessionListRowContextProvider;
-  defaultLimit?: number;
-}): SessionEntrySelection {
-  const creatorEntries = filterSessionEntries(params);
-  const creatorId = normalizeOptionalString(params.opts.creatorId);
-  const filtered = creatorId
-    ? creatorEntries.filter(([, entry]) => entry.createdActor?.id === creatorId)
-    : creatorEntries;
-  const limit = resolveSessionsListLimit(params.opts, params.defaultLimit);
-  const offset = resolveSessionsListOffset(params.opts);
-  const windowLimit = resolveSessionsListWindowLimit(limit, offset);
-  const sortedWindow = sortAndLimitSessionEntries(filtered, windowLimit, params.opts.sortBy);
-  const entries =
-    limit === undefined ? sortedWindow.slice(offset) : sortedWindow.slice(offset, offset + limit);
-  const nextOffset = offset + entries.length;
-  const hasMore = nextOffset < filtered.length;
-  return {
-    entries,
-    creatorEntries,
-    totalCount: filtered.length,
-    limitApplied: limit,
-    offset,
-    nextOffset: hasMore ? nextOffset : null,
-    hasMore,
-  };
-}
-
-function listSessionCreatorIdentities(
-  entries: readonly SessionEntryPair[],
-  userProfileLabelById: Map<string, string | undefined>,
-): Array<{ id: string; label?: string }> {
-  const creators = new Map<string, { id: string; label?: string }>();
-  for (const [, entry] of entries) {
-    const actor = projectSessionActor(entry.createdActor, userProfileLabelById);
-    const id = normalizeOptionalString(actor?.id);
-    if (!id) {
-      continue;
-    }
-    const label = normalizeOptionalString(actor?.label);
-    const existing = creators.get(id);
-    if (!existing || (label && (!existing.label || label.localeCompare(existing.label) < 0))) {
-      creators.set(id, { id, ...(label ? { label } : {}) });
-    }
-  }
-  return [...creators.values()].toSorted((a, b) => {
-    const byLabel = (a.label ?? a.id).localeCompare(b.label ?? b.id);
-    return byLabel || a.id.localeCompare(b.id);
-  });
-}
-
-export function filterAndSortSessionEntries(params: {
-  cfg: OpenClawConfig;
-  store: Record<string, SessionEntry>;
-  opts: SessionsListParams;
-  now: number;
-  rowContext?: SessionListRowContext;
-  getRowContext?: SessionListRowContextProvider;
-}): [string, SessionEntry][] {
-  return selectSessionEntries(params).entries;
-}
-
-export function listSessionsFromStore(params: {
-  cfg: OpenClawConfig;
-  storePath: string;
-  store: Record<string, SessionEntry>;
-  modelCatalog?: ModelCatalogEntry[];
-  opts: SessionsListParams;
-}): SessionsListResult {
-  const { cfg, storePath, store, opts } = params;
-  const now = Date.now();
-  const sessionListTranscriptUsageMaxBytes = 64 * 1024;
-  const sessionListTranscriptFieldRows = 100;
-  let rowContext: SessionListRowContext | undefined;
-  const getRowContext = () => {
-    rowContext ??= buildSessionListRowContext({ store, now });
-    return rowContext;
-  };
-  const includeDerivedTitles = opts.includeDerivedTitles === true;
-  const includeLastMessage = opts.includeLastMessage === true;
-  const hasSpawnedByFilter = typeof opts.spawnedBy === "string" && opts.spawnedBy.length > 0;
-
-  const selection = selectSessionEntries({
-    cfg,
-    store,
-    opts,
-    now,
-    getRowContext:
-      hasSpawnedByFilter || Boolean(normalizeOptionalString(opts.search))
-        ? getRowContext
-        : undefined,
-    defaultLimit: SESSIONS_LIST_DEFAULT_LIMIT,
-  });
-  const { entries, creatorEntries, totalCount, limitApplied, offset, nextOffset, hasMore } =
-    selection;
-  const fullRowContext =
-    rowContext || hasSpawnedByFilter || entries.length > SESSIONS_LIST_YIELD_BATCH_SIZE
-      ? getRowContext()
-      : undefined;
-  const sharedRowContext =
-    fullRowContext ??
-    (entries.length > 0 ? buildSessionListRowMetadataContext({ now }) : undefined);
-
-  const sessions = entries.map(([key, entry], index) => {
-    const includeTranscriptFields = index < sessionListTranscriptFieldRows;
-    const rowAgentId =
-      key === "global" && typeof opts.agentId === "string"
-        ? normalizeAgentId(opts.agentId)
-        : undefined;
-    const storeChildSessionsByKey =
-      fullRowContext?.storeChildSessionsByKey ??
-      buildSingleRowStoreChildSessionsByKey({ store, storePath, key, now });
-    return buildGatewaySessionRow({
-      cfg,
-      storePath,
-      store,
-      key,
-      entry,
-      agentId: rowAgentId,
-      modelCatalog: params.modelCatalog,
-      now,
-      includeDerivedTitles: includeTranscriptFields && includeDerivedTitles,
-      includeLastMessage: includeTranscriptFields && includeLastMessage,
-      transcriptUsageMaxBytes: sessionListTranscriptUsageMaxBytes,
-      storeChildSessionsByKey,
-      rowContext: sharedRowContext,
-    });
-  });
-
-  return {
-    ts: now,
-    path: storePath,
-    count: sessions.length,
-    totalCount,
-    limitApplied,
-    offset: offset > 0 ? offset : undefined,
-    nextOffset,
-    hasMore,
-    creators: listSessionCreatorIdentities(
-      creatorEntries,
-      sharedRowContext?.userProfileLabelById ?? new Map(),
-    ),
-    defaults: getSessionDefaults(cfg, params.modelCatalog, { allowPluginNormalization: false }),
-    sessions,
-  };
-}
-
-/**
- * Async version of listSessionsFromStore that yields to the event loop between
- * batches of session row builds. This prevents large session stores from
- * blocking the event loop during sessions.list requests.
- *
- * The synchronous file I/O in readSessionTitleFieldsFromTranscript (head/tail
- * reads for derived titles and last-message previews) is the dominant blocker.
- * By yielding every SESSIONS_LIST_YIELD_BATCH_SIZE rows, we keep the event
- * loop responsive for WebSocket heartbeats, channel I/O, and concurrent RPC.
- */
-export async function listSessionsFromStoreAsync(params: {
-  cfg: OpenClawConfig;
-  storePath: string;
-  store: Record<string, SessionEntry>;
-  modelCatalog?: ModelCatalogEntry[];
-  opts: SessionsListParams;
-}): Promise<SessionsListResult> {
-  // Pin the active plugin-registry workspace dir for the duration of this
-  // call so per-row metadata lookups use a stable memo key. Without this pin,
-  // concurrent agent turns / crons mutate the process-global workspace dir
-  // between rows, the memo never hits, and each row triggers a full
-  // loadPluginMetadataSnapshot scan (~100 ms).
-  return withPinnedActivePluginRegistryWorkspaceDir(async () => {
-    const { cfg, storePath, store, opts } = params;
-    const now = Date.now();
-    const sessionListTranscriptUsageMaxBytes = 64 * 1024;
-    const sessionListTranscriptFieldRows = 100;
-    let rowContext: SessionListRowContext | undefined;
-    const getRowContext = () => {
-      rowContext ??= buildSessionListRowContext({ store, now });
-      return rowContext;
-    };
-    const includeDerivedTitles = opts.includeDerivedTitles === true;
-    const includeLastMessage = opts.includeLastMessage === true;
-    const hasSpawnedByFilter = typeof opts.spawnedBy === "string" && opts.spawnedBy.length > 0;
-
-    const selection = selectSessionEntries({
-      cfg,
-      store,
-      opts,
-      now,
-      getRowContext:
-        hasSpawnedByFilter || Boolean(normalizeOptionalString(opts.search))
-          ? getRowContext
-          : undefined,
-      defaultLimit: SESSIONS_LIST_DEFAULT_LIMIT,
-    });
-    const { entries, creatorEntries, totalCount, limitApplied, offset, nextOffset, hasMore } =
-      selection;
-    const fullRowContext =
-      rowContext || hasSpawnedByFilter || entries.length > SESSIONS_LIST_YIELD_BATCH_SIZE
-        ? getRowContext()
-        : undefined;
-    const sharedRowContext =
-      fullRowContext ??
-      (entries.length > 0 ? buildSessionListRowMetadataContext({ now }) : undefined);
-
-    const sessions: GatewaySessionRow[] = [];
-    for (let i = 0; i < entries.length; i++) {
-      const [key, entry] = expectDefined(entries[i], "entries entry at i");
-      const includeTranscriptFields = i < sessionListTranscriptFieldRows;
-      const rowAgentId =
-        key === "global" && typeof opts.agentId === "string"
-          ? normalizeAgentId(opts.agentId)
-          : undefined;
-      const storeChildSessionsByKey =
-        fullRowContext?.storeChildSessionsByKey ??
-        buildSingleRowStoreChildSessionsByKey({ store, storePath, key, now });
-      const row = buildGatewaySessionRow({
-        cfg,
-        storePath,
-        store,
-        key,
-        entry,
-        agentId: rowAgentId,
-        modelCatalog: params.modelCatalog,
-        now,
-        includeDerivedTitles: false,
-        includeLastMessage: false,
-        transcriptUsageMaxBytes: sessionListTranscriptUsageMaxBytes,
-        storeChildSessionsByKey,
-        rowContext: sharedRowContext,
-        skipTranscriptUsageFallback: true,
-        lightweightListRow: true,
-      });
-      if (
-        entry?.sessionId &&
-        includeTranscriptFields &&
-        (includeDerivedTitles || includeLastMessage)
-      ) {
-        const parsed = parseAgentSessionKey(key);
-        const sessionAgentId =
-          rowAgentId ??
-          (parsed?.agentId ? normalizeAgentId(parsed.agentId) : resolveDefaultAgentId(cfg));
-        const fields = await readScopedSessionTitleFieldsFromTranscriptAsync({
-          agentId: sessionAgentId,
-          sessionEntry: entry,
-          sessionId: entry.sessionId,
-          sessionKey: key,
-          storePath,
-        });
-        if (includeDerivedTitles) {
-          row.derivedTitle = deriveSessionTitle(entry, fields.firstUserMessage);
-        }
-        if (includeLastMessage && fields.lastMessagePreview) {
-          row.lastMessagePreview = fields.lastMessagePreview;
-        }
-      }
-      sessions.push(row);
-      // Yield to the event loop between batches so WebSocket heartbeats,
-      // channel I/O, and concurrent RPC calls are not starved.
-      if ((i + 1) % SESSIONS_LIST_YIELD_BATCH_SIZE === 0 && i + 1 < entries.length) {
-        await new Promise<void>((resolve) => {
-          setImmediate(resolve);
-        });
-      }
-    }
-
-    return {
-      ts: now,
-      path: storePath,
-      count: sessions.length,
-      totalCount,
-      limitApplied,
-      offset: offset > 0 ? offset : undefined,
-      nextOffset,
-      hasMore,
-      creators: listSessionCreatorIdentities(
-        creatorEntries,
-        sharedRowContext?.userProfileLabelById ?? new Map(),
-      ),
-      defaults: getSessionDefaults(cfg, params.modelCatalog, { allowPluginNormalization: false }),
-      sessions,
-    };
-  });
-}
-/* oxlint-disable max-lines -- TODO: split this grandfathered oversized file. */
+export { deriveSessionTitle } from "./session-utils-core.js";
+export { resolveDeletedAgentIdFromSessionKey } from "./session-utils-store.js";
+export { loadSessionEntry } from "./session-utils-store.js";
+export { loadSessionEntryReadOnly } from "./session-utils-store.js";
+export { resolveFreshestSessionStoreMatchFromStoreKeys } from "./session-utils-store.js";
+export { resolveFreshestSessionEntryFromStoreKeys } from "./session-utils-store.js";
+export { migrateAndPruneGatewaySessionStoreKey } from "./session-utils-store.js";
+export { listAgentsForGateway } from "./session-utils-store.js";
+export { resolveGatewaySessionStoreTargetWithStore } from "./session-utils-store-lookup.js";
+export { resolveGatewaySessionStoreTarget } from "./session-utils-store-lookup.js";
+export { resolveGatewaySessionThinkingProjection } from "./session-utils-model.js";
+export { getSessionDefaults } from "./session-utils-model.js";
+export { resolveGatewayModelSupportsImages } from "./session-utils-model.js";
+export { resolveSessionDisplayModelIdentityRef } from "./session-utils-model.js";
+export { buildGatewaySessionRow } from "./session-utils-row.js";
+export { loadGatewaySessionRow } from "./session-utils-search.js";
+export { buildGatewaySessionInfo } from "./session-utils-search.js";
+export { filterAndSortSessionEntries } from "./session-utils-list.js";
+export { listSessionsFromStore } from "./session-utils-list.js";
+export { listSessionsFromStoreAsync } from "./session-utils-list.js";
